+++ pkg/pipeline/PipelineBuilder.go
@@ -21,54 +21,47 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/devtron-labs/devtron/internal/sql/repository/helper"
 	"github.com/devtron-labs/devtron/pkg/variables"
 	"github.com/devtron-labs/devtron/pkg/variables/parsers"
 	repository6 "github.com/devtron-labs/devtron/pkg/variables/repository"
 	"net/http"
 	"net/url"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
 
 	application2 "github.com/argoproj/argo-cd/v2/pkg/apiclient/application"
-	client "github.com/devtron-labs/devtron/api/helm-app"
-	app2 "github.com/devtron-labs/devtron/internal/sql/repository/app"
-	"github.com/devtron-labs/devtron/internal/sql/repository/appStatus"
-	"github.com/devtron-labs/devtron/internal/sql/repository/helper"
-	"github.com/devtron-labs/devtron/internal/sql/repository/security"
-	"github.com/devtron-labs/devtron/pkg/chart"
-	chartRepoRepository "github.com/devtron-labs/devtron/pkg/chartRepo/repository"
-	"github.com/devtron-labs/devtron/pkg/cluster"
-	repository2 "github.com/devtron-labs/devtron/pkg/cluster/repository"
-	bean3 "github.com/devtron-labs/devtron/pkg/pipeline/bean"
-	"github.com/devtron-labs/devtron/pkg/pipeline/history"
-	repository4 "github.com/devtron-labs/devtron/pkg/pipeline/history/repository"
-	repository5 "github.com/devtron-labs/devtron/pkg/pipeline/repository"
-	resourceGroup2 "github.com/devtron-labs/devtron/pkg/resourceGroup"
-	"github.com/devtron-labs/devtron/pkg/sql"
-	"github.com/devtron-labs/devtron/pkg/user"
-	util3 "github.com/devtron-labs/devtron/pkg/util"
-	"github.com/devtron-labs/devtron/util/argo"
-	util4 "github.com/devtron-labs/devtron/util/k8s"
-	"github.com/devtron-labs/devtron/util/rbac"
-	"go.opentelemetry.io/otel"
-
 	"github.com/caarlos0/env"
 	bean2 "github.com/devtron-labs/devtron/api/bean"
+	client "github.com/devtron-labs/devtron/api/helm-app"
 	"github.com/devtron-labs/devtron/client/argocdServer"
 	"github.com/devtron-labs/devtron/client/argocdServer/application"
 	"github.com/devtron-labs/devtron/internal/sql/models"
 	"github.com/devtron-labs/devtron/internal/sql/repository"
+	app2 "github.com/devtron-labs/devtron/internal/sql/repository/app"
+	"github.com/devtron-labs/devtron/internal/sql/repository/appStatus"
 	"github.com/devtron-labs/devtron/internal/sql/repository/appWorkflow"
 	"github.com/devtron-labs/devtron/internal/sql/repository/chartConfig"
 	dockerRegistryRepository "github.com/devtron-labs/devtron/internal/sql/repository/dockerRegistry"
 	"github.com/devtron-labs/devtron/internal/sql/repository/pipelineConfig"
+	"github.com/devtron-labs/devtron/internal/sql/repository/security"
 	"github.com/devtron-labs/devtron/internal/util"
 	"github.com/devtron-labs/devtron/pkg/app"
 	"github.com/devtron-labs/devtron/pkg/attributes"
 	"github.com/devtron-labs/devtron/pkg/bean"
+	"github.com/devtron-labs/devtron/pkg/chart"
+	chartRepoRepository "github.com/devtron-labs/devtron/pkg/chartRepo/repository"
+	repository2 "github.com/devtron-labs/devtron/pkg/cluster/repository"
+	"github.com/devtron-labs/devtron/pkg/pipeline/history"
+	resourceGroup2 "github.com/devtron-labs/devtron/pkg/resourceGroup"
+	"github.com/devtron-labs/devtron/pkg/sql"
+	"github.com/devtron-labs/devtron/pkg/user"
+	util3 "github.com/devtron-labs/devtron/pkg/util"
 	util2 "github.com/devtron-labs/devtron/util"
+	"github.com/devtron-labs/devtron/util/argo"
+	util4 "github.com/devtron-labs/devtron/util/k8s"
+	"github.com/devtron-labs/devtron/util/rbac"
 	"github.com/go-pg/pg"
 	"github.com/juju/errors"
 	"go.uber.org/zap"
@@ -105,9 +98,7 @@ func GetDeploymentServiceTypeConfig() (*DeploymentServiceTypeConfig, error) {
 	return cfg, err
 }
 
-type PipelineBuilder interface {
-	//function related to application and project
-
+type DevtronAppService interface {
 	//CreateApp : This function creates applications of type Job as well as Devtronapps
 	// In case of error response object is nil
 	CreateApp(request *bean.CreateAppDTO) (*bean.CreateAppDTO, error)
@@ -149,168 +140,37 @@ type PipelineBuilder interface {
 	// The function returns the list of applications in valid case.
 	// In case of error,[]*AppBean is returned as nil.
 	FindAppsByTeamName(teamName string) ([]AppBean, error)
+}
 
+type PipelineBuilder interface {
+	//function related to application and project
+
+	DevtronAppService
 	//function related to ciPipeline
 
-	//GetCiPipeline : retrieves CI pipeline configuration (CiConfigRequest) for a specific application (appId).
-	// It fetches CI pipeline data, including pipeline materials, scripts, and associated configurations.
-	// It returns a detailed CiConfigRequest.
-	//If any errors occur during the retrieval process  CI pipeline configuration remains nil.
-	//If you want less detail of ciPipeline ,Please refer GetCiPipelineMin
-	GetCiPipeline(appId int) (ciConfig *bean.CiConfigRequest, err error)
-	//GetCiPipelineById : Retrieve ciPipeline for given ciPipelineId
-	GetCiPipelineById(pipelineId int) (ciPipeline *bean.CiPipeline, err error)
-	//GetTriggerViewCiPipeline : retrieves a detailed view of the CI pipelines configured for a specific application (appId).
-	// It includes information on CI pipeline materials, scripts, configurations, and linked pipelines.
-	//If any errors occur ,It returns an error along with a nil result(bean.TriggerViewCiConfig).
-	GetTriggerViewCiPipeline(appId int) (*bean.TriggerViewCiConfig, error)
-	//GetExternalCi : Lists externalCi for given appId
-	//"external CI" refers to CI pipelines and configurations that are managed externally,
-	//like by third-party services or tools.
-	//It fetches information about external CI pipelines, their webhooks, payload configurations, and related roles.
-	//The function constructs an array of ExternalCiConfig objects and returns it.
-	// If any errors occur, the function returns an error along with a nil result([]*bean.ExternalCiConfig).
-	GetExternalCi(appId int) (ciConfig []*bean.ExternalCiConfig, err error)
-	//GetExternalCiById : Retrieve externalCi for given appId and externalCiId.
-	//It begins by validating the provided ID and fetching the corresponding CI pipeline from the repository.
-	//If the CI pipeline is found, the function constructs an ExternalCiConfig object, encapsulating essential details and returns it.
-	// If any errors occur, the function returns an error along with a nil result(*bean.ExternalCiConfig).
-	GetExternalCiById(appId int, externalCiId int) (ciConfig *bean.ExternalCiConfig, err error)
-	//UpdateCiTemplate : handles updates to the CiTemplate based on the provided CiConfigRequest.
-	//It fetches relevant Docker registry information, and updates the CI configuration.
-	//The function then creates or modifies associated resources such as Docker repositories.
-	//After updating the configuration, it ensures to update history .
-	//Finally, the function returns the modified CiConfigRequest.
-	//If an error occurs, the CiConfigRequest is nil.
-	//If you want to Update CiPipeline please refer PatchCiPipeline
-	UpdateCiTemplate(updateRequest *bean.CiConfigRequest) (*bean.CiConfigRequest, error)
-	//PatchCiPipeline :  function manages CI pipeline operations based on the provided CiPatchRequest.
-	//It fetches template variables, sets specific attributes, and
-	//handles following actions
-	// 1. create
-	//2. update source
-	//3. delete pipelines
-	PatchCiPipeline(request *bean.CiPatchRequest) (ciConfig *bean.CiConfigRequest, err error)
-	//CreateCiPipeline : manages the creation of a CI pipeline based on the provided CiConfigRequest.
-	// It first fetches  application data and configures Docker registry settings
-	//then constructs a CI template with specified build configurations, auditing details, and related Git materials.
-
-	CreateCiPipeline(createRequest *bean.CiConfigRequest) (*bean.PipelineCreateResponse, error)
-	//GetCiPipelineMin : lists minimum detail of ciPipelines for given appId and envIds.
-	//It filters and fetches CI pipelines based on the provided environment identifiers.
-	//If no specific environments are provided, it retrieves all CI pipelines associated with the application.
-	//If you want more details like buildConfig ,gitMaterials etc, please refer GetCiPipeline
-	GetCiPipelineMin(appId int, envIds []int) ([]*bean.CiPipelineMin, error)
-	//PatchRegexCiPipeline : Update CI pipeline materials based on the provided regex patch request
-	PatchRegexCiPipeline(request *bean.CiRegexPatchRequest) (err error)
-	//GetCiPipelineByEnvironment : lists ciPipeline for given environmentId and appIds
-	GetCiPipelineByEnvironment(request resourceGroup2.ResourceGroupingRequest) ([]*bean.CiConfigRequest, error)
-	//GetCiPipelineByEnvironmentMin : lists minimum detail of ciPipelines for given environmentId and appIds
-	GetCiPipelineByEnvironmentMin(request resourceGroup2.ResourceGroupingRequest) ([]*bean.CiPipelineMinResponse, error)
-	//GetExternalCiByEnvironment : lists externalCi for given environmentId and appIds
-	GetExternalCiByEnvironment(request resourceGroup2.ResourceGroupingRequest) (ciConfig []*bean.ExternalCiConfig, err error)
-	DeleteCiPipeline(request *bean.CiPatchRequest) (*bean.CiPipeline, error)
-
+	DevtronAppBuildService
 	//function related to gitMaterial
 
-	//CreateMaterialsForApp : Delegating the request to ciCdPipelineOrchestrator for Material creation
-	CreateMaterialsForApp(request *bean.CreateMaterialDTO) (*bean.CreateMaterialDTO, error)
-	//UpdateMaterialsForApp : Delegating the request to ciCdPipelineOrchestrator for updating Material
-	UpdateMaterialsForApp(request *bean.UpdateMaterialDTO) (*bean.UpdateMaterialDTO, error)
-	DeleteMaterial(request *bean.UpdateMaterialDTO) error
-	//PatchCiMaterialSource : Delegating the request to ciCdPipelineOrchestrator for updating source
-	PatchCiMaterialSource(ciPipeline *bean.CiMaterialPatchRequest, userId int32) (*bean.CiMaterialPatchRequest, error)
-	//BulkPatchCiMaterialSource : Delegating the request to ciCdPipelineOrchestrator for bulk updating source
-	BulkPatchCiMaterialSource(ciPipelines *bean.CiMaterialBulkPatchRequest, userId int32, token string, checkAppSpecificAccess func(token, action string, appId int) (bool, error)) (*bean.CiMaterialBulkPatchResponse, error)
-	//GetMaterialsForAppId : Retrieve material for given appId
-	GetMaterialsForAppId(appId int) []*bean.GitMaterial
+	DevtronAppMaterialBuilderService
 
 	//function related to cdPipeline
 
-	//GetCdPipelineById : Retrieve cdPipeline for given cdPipelineId.
-	//getting cdPipeline,environment and strategies ,preDeployStage, postDeployStage,appWorkflowMapping from respective repository and service layer
-	//converting above data in proper bean object and then assigning to CDPipelineConfigObject
-	//if any error occur , will get empty object or nil
-	GetCdPipelineById(pipelineId int) (cdPipeline *bean.CDPipelineConfigObject, err error)
-	CreateCdPipelines(cdPipelines *bean.CdPipelines, ctx context.Context) (*bean.CdPipelines, error)
-	//PatchCdPipelines : Handle CD pipeline patch requests, making necessary changes to the configuration and returning the updated version.
-	//Performs Create ,Update and Delete operation.
-	PatchCdPipelines(cdPipelines *bean.CDPatchRequest, ctx context.Context) (*bean.CdPipelines, error)
-	DeleteCdPipeline(pipeline *pipelineConfig.Pipeline, ctx context.Context, deleteAction int, acdDelete bool, userId int32) (*bean.AppDeleteResponseDTO, error)
-	DeleteACDAppCdPipelineWithNonCascade(pipeline *pipelineConfig.Pipeline, ctx context.Context, forceDelete bool, userId int32) (err error)
-	//GetTriggerViewCdPipelinesForApp :
-	GetTriggerViewCdPipelinesForApp(appId int) (cdPipelines *bean.CdPipelines, err error)
-	//GetCdPipelinesForApp : Retrieve cdPipeline for given appId
-	GetCdPipelinesForApp(appId int) (cdPipelines *bean.CdPipelines, err error)
-	//GetCdPipelinesForAppAndEnv : Retrieve cdPipeline for given appId and envId
-	GetCdPipelinesForAppAndEnv(appId int, envId int) (cdPipelines *bean.CdPipelines, err error)
-	/*	CreateCdPipelines(cdPipelines bean.CdPipelines) (*bean.CdPipelines, error)*/
-	//GetCdPipelinesByEnvironment : lists cdPipeline for given environmentId and appIds
-	GetCdPipelinesByEnvironment(request resourceGroup2.ResourceGroupingRequest) (cdPipelines *bean.CdPipelines, err error)
-	//GetCdPipelinesByEnvironmentMin : lists minimum detail of cdPipelines for given environmentId and appIds
-	GetCdPipelinesByEnvironmentMin(request resourceGroup2.ResourceGroupingRequest) (cdPipelines []*bean.CDPipelineConfigObject, err error)
-	//PerformBulkActionOnCdPipelines :
-	PerformBulkActionOnCdPipelines(dto *bean.CdBulkActionRequestDto, impactedPipelines []*pipelineConfig.Pipeline, ctx context.Context, dryRun bool, userId int32) ([]*bean.CdBulkActionResponseDto, error)
-	//FindPipelineById : Retrieve Pipeline object from pipelineRepository for given cdPipelineId
-	FindPipelineById(cdPipelineId int) (*pipelineConfig.Pipeline, error)
-	//FindAppAndEnvDetailsByPipelineId : Retrieve app and env details for given cdPipelineId
-	FindAppAndEnvDetailsByPipelineId(cdPipelineId int) (*pipelineConfig.Pipeline, error)
-	// RetrieveParentDetails : Retrieve the parent id and type of the parent.
-	//Here ParentId refers to Parent like parent of CD can be CI , PRE-CD .
-	// It first fetches the workflow details from the appWorkflow repository.
-	//If the workflow is a CD pipeline, it further checks for stage configurations.
-	//If the workflow is a webhook, it returns the webhook workflow type.
-	//In case of error , it returns 0 for parentId and empty string for parentType
-	RetrieveParentDetails(pipelineId int) (parentId int, parentType bean2.WorkflowType, err error)
-	//GetEnvironmentByCdPipelineId : Retrieve environmentId for given cdPipelineId
-	GetEnvironmentByCdPipelineId(pipelineId int) (int, error)
-	GetBulkActionImpactedPipelines(dto *bean.CdBulkActionRequestDto) ([]*pipelineConfig.Pipeline, error) //no usage
-	//IsGitOpsRequiredForCD : Determine if GitOps is required for CD based on the provided pipeline creation request
-	IsGitOpsRequiredForCD(pipelineCreateRequest *bean.CdPipelines) bool
-	//SetPipelineDeploymentAppType : Set pipeline deployment application(helm/argo) types based on the provided configuration
-	SetPipelineDeploymentAppType(pipelineCreateRequest *bean.CdPipelines, isGitOpsConfigured bool, deploymentTypeValidationConfig map[string]bool)
-	MarkGitOpsDevtronAppsDeletedWhereArgoAppIsDeleted(appId int, envId int, acdToken string, pipeline *pipelineConfig.Pipeline) (bool, error)
-	//GetEnvironmentListForAutocompleteFilter : lists environment for given configuration
-	GetEnvironmentListForAutocompleteFilter(envName string, clusterIds []int, offset int, size int, emailId string, checkAuthBatch func(emailId string, appObject []string, envObject []string) (map[string]bool, map[string]bool), ctx context.Context) (*cluster.ResourceGroupingResponse, error)
-	IsGitopsConfigured() (bool, error)
+	DevtronAppDeploymentService
 
 	//function related to CM/CS
 
-	//FetchConfigmapSecretsForCdStages : Delegating the request to appService for fetching cm/cs
-	FetchConfigmapSecretsForCdStages(appId, envId, cdPipelineId int) (ConfigMapSecretsResponse, error)
-	//GetDeploymentConfigMap : Retrieve deployment config values from the attributes table
-	GetDeploymentConfigMap(environmentId int) (map[string]bool, error)
+	DevtronAppCMCSService
 
 	//function related to strategy
 
-	//FetchCDPipelineStrategy : Retrieve CDPipelineStrategy for given appId
-	FetchCDPipelineStrategy(appId int) (PipelineStrategiesResponse, error)
-	//FetchDefaultCDPipelineStrategy :
-	FetchDefaultCDPipelineStrategy(appId int, envId int) (PipelineStrategy, error)
+	DevtronAppStrategyService
 
 	//function related to DeploymentAppTypeChangeResponse
 
-	//ChangeDeploymentType : takes in DeploymentAppTypeChangeRequest struct and
-	// deletes all the cd pipelines for that deployment type in all apps that belongs to
-	// that environment and updates the db with desired deployment app type
-	ChangeDeploymentType(ctx context.Context, request *bean.DeploymentAppTypeChangeRequest) (*bean.DeploymentAppTypeChangeResponse, error)
-	//ChangePipelineDeploymentType : takes in DeploymentAppTypeChangeRequest struct and
-	// deletes all the cd pipelines for that deployment type in all apps that belongs to
-	// that environment and updates the db with desired deployment app type
-	ChangePipelineDeploymentType(ctx context.Context, request *bean.DeploymentAppTypeChangeRequest) (*bean.DeploymentAppTypeChangeResponse, error)
-	//TriggerDeploymentAfterTypeChange :
-	TriggerDeploymentAfterTypeChange(ctx context.Context, request *bean.DeploymentAppTypeChangeRequest) (*bean.DeploymentAppTypeChangeResponse, error)
-	//DeleteDeploymentApps :
-	DeleteDeploymentApps(ctx context.Context, pipelines []*pipelineConfig.Pipeline, userId int32) *bean.DeploymentAppTypeChangeResponse
-	//DeleteDeploymentAppsForEnvironment :
-	DeleteDeploymentAppsForEnvironment(ctx context.Context, environmentId int, currentDeploymentAppType bean.DeploymentType, exclusionList []int, includeApps []int, userId int32) (*bean.DeploymentAppTypeChangeResponse, error)
-
+	DevtronAppDeploymentTypeChangeService
 	//function related to artifact
 
-	//RetrieveArtifactsByCDPipeline :
-	RetrieveArtifactsByCDPipeline(pipeline *pipelineConfig.Pipeline, stage bean2.WorkflowType) (*bean.CiArtifactResponse, error)
-	//FetchArtifactForRollback :
-	FetchArtifactForRollback(cdPipelineId, appId, offset, limit int) (bean.CiArtifactResponse, error)
+	DevtronAppArtifactService
 }
 type PipelineBuilderImpl struct {
 	logger                        *zap.SugaredLogger
@@ -526,6 +386,12 @@ func formatDate(t time.Time, layout string) string {
 	return t.Format(layout)
 }
 
+/*
+   1. create pipelineGroup
+   2. save material (add credential provider support)
+
+*/
+
 func (impl *PipelineBuilderImpl) CreateApp(request *bean.CreateAppDTO) (*bean.CreateAppDTO, error) {
 	impl.logger.Debugw("app create request received", "req", request)
 
@@ -540,93 +406,6 @@ func (impl *PipelineBuilderImpl) DeleteApp(appId int, userId int32) error {
 	impl.logger.Debugw("app delete request received", "app", appId)
 	err := impl.ciCdPipelineOrchestrator.DeleteApp(appId, userId)
 	return err
-}
-
-func (impl *PipelineBuilderImpl) CreateMaterialsForApp(request *bean.CreateMaterialDTO) (*bean.CreateMaterialDTO, error) {
-	res, err := impl.ciCdPipelineOrchestrator.CreateMaterials(request)
-	if err != nil {
-		impl.logger.Errorw("error in saving create materials req", "req", request, "err", err)
-	}
-	return res, err
-}
-
-func (impl *PipelineBuilderImpl) UpdateMaterialsForApp(request *bean.UpdateMaterialDTO) (*bean.UpdateMaterialDTO, error) {
-	res, err := impl.ciCdPipelineOrchestrator.UpdateMaterial(request)
-	if err != nil {
-		impl.logger.Errorw("error in updating materials req", "req", request, "err", err)
-	}
-	return res, err
-}
-
-func (impl *PipelineBuilderImpl) DeleteMaterial(request *bean.UpdateMaterialDTO) error {
-	//finding ci pipelines for this app; if found any, will not delete git material
-	pipelines, err := impl.ciPipelineRepository.FindByAppId(request.AppId)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("err in deleting git material", "gitMaterial", request.Material, "err", err)
-		return err
-	}
-	if len(pipelines) > 0 {
-		//pipelines are present, in this case we will check if this material is used in docker config
-		//if it is used, then we won't delete
-		ciTemplateBean, err := impl.ciTemplateService.FindByAppId(request.AppId)
-		if err != nil && err == errors.NotFoundf(err.Error()) {
-			impl.logger.Errorw("err in getting docker registry", "appId", request.AppId, "err", err)
-			return err
-		}
-		if ciTemplateBean != nil {
-			ciTemplate := ciTemplateBean.CiTemplate
-			if ciTemplate != nil && ciTemplate.GitMaterialId == request.Material.Id {
-				return fmt.Errorf("cannot delete git material, is being used in docker config")
-			}
-		}
-	}
-	existingMaterial, err := impl.materialRepo.FindById(request.Material.Id)
-	if err != nil {
-		impl.logger.Errorw("No matching entry found for delete", "gitMaterial", request.Material)
-		return err
-	}
-	existingMaterial.UpdatedOn = time.Now()
-	existingMaterial.UpdatedBy = request.UserId
-
-	err = impl.materialRepo.MarkMaterialDeleted(existingMaterial)
-	if err != nil {
-		impl.logger.Errorw("error in deleting git material", "gitMaterial", existingMaterial)
-		return err
-	}
-
-	err = impl.gitMaterialHistoryService.MarkMaterialDeletedAndCreateHistory(existingMaterial)
-
-	dbConnection := impl.pipelineRepository.GetConnection()
-	tx, err := dbConnection.Begin()
-	if err != nil {
-		return err
-	}
-	// Rollback tx on error.
-	defer tx.Rollback()
-	var materials []*pipelineConfig.CiPipelineMaterial
-	for _, pipeline := range pipelines {
-		materialDbObject, err := impl.ciPipelineMaterialRepository.GetByPipelineIdAndGitMaterialId(pipeline.Id, request.Material.Id)
-		if err != nil {
-			return err
-		}
-		if len(materialDbObject) == 0 {
-			continue
-		}
-		materialDbObject[0].Active = false
-		materials = append(materials, materialDbObject[0])
-	}
-
-	if len(materials) == 0 {
-		return nil
-	}
-
-	err = impl.ciPipelineMaterialRepository.Update(tx, materials...)
-	err = tx.Commit()
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func (impl *PipelineBuilderImpl) GetApp(appId int) (application *bean.CreateAppDTO, err error) {
@@ -653,45 +432,124 @@ func (impl *PipelineBuilderImpl) GetApp(appId int) (application *bean.CreateAppD
 	return application, nil
 }
 
-func (impl *PipelineBuilderImpl) GetMaterialsForAppId(appId int) []*bean.GitMaterial {
-	materials, err := impl.materialRepo.FindByAppId(appId)
+func (impl *PipelineBuilderImpl) FindByIds(ids []*int) ([]*AppBean, error) {
+	var appsRes []*AppBean
+	apps, err := impl.appRepo.FindByIds(ids)
 	if err != nil {
-		impl.logger.Errorw("error in fetching materials", "appId", appId, "err", err)
+		impl.logger.Errorw("error while fetching app", "err", err)
+		return nil, err
 	}
-
-	ciTemplateBean, err := impl.ciTemplateService.FindByAppId(appId)
-	if err != nil && err != errors.NotFoundf(err.Error()) {
-		impl.logger.Errorw("err in getting ci-template", "appId", appId, "err", err)
+	for _, app := range apps {
+		appsRes = append(appsRes, &AppBean{Id: app.Id, Name: app.AppName, TeamId: app.TeamId})
 	}
-
-	var gitMaterials []*bean.GitMaterial
-	for _, material := range materials {
-		gitMaterial := &bean.GitMaterial{
-			Url:             material.Url,
-			Name:            material.Name[strings.Index(material.Name, "-")+1:],
-			Id:              material.Id,
-			GitProviderId:   material.GitProviderId,
-			CheckoutPath:    material.CheckoutPath,
-			FetchSubmodules: material.FetchSubmodules,
-			FilterPattern:   material.FilterPattern,
-		}
-		//check if git material is deletable or not
-		if ciTemplateBean != nil {
-			ciTemplate := ciTemplateBean.CiTemplate
-			if ciTemplate != nil && (ciTemplate.GitMaterialId == material.Id || ciTemplate.BuildContextGitMaterialId == material.Id) {
-				gitMaterial.IsUsedInCiConfig = true
-			}
-		}
-		gitMaterials = append(gitMaterials, gitMaterial)
-	}
-	return gitMaterials
+	return appsRes, err
 }
 
-/*
-   1. create pipelineGroup
-   2. save material (add credential provider support)
+func (impl *PipelineBuilderImpl) GetAppList() ([]AppBean, error) {
+	var appsRes []AppBean
+	apps, err := impl.appRepo.FindAll()
+	if err != nil {
+		impl.logger.Errorw("error while fetching app", "err", err)
+		return nil, err
+	}
+	for _, app := range apps {
+		appsRes = append(appsRes, AppBean{Id: app.Id, Name: app.AppName})
+	}
+	return appsRes, err
+}
 
-*/
+func (impl *PipelineBuilderImpl) FindAllMatchesByAppName(appName string, appType helper.AppType) ([]*AppBean, error) {
+	var appsRes []*AppBean
+	var apps []*app2.App
+	var err error
+	if len(appName) == 0 {
+		apps, err = impl.appRepo.FindAll()
+	} else {
+		apps, err = impl.appRepo.FindAllMatchesByAppName(appName, appType)
+	}
+	if err != nil {
+		impl.logger.Errorw("error while fetching app", "err", err)
+		return nil, err
+	}
+	for _, app := range apps {
+		name := app.AppName
+		if appType == helper.Job {
+			name = app.DisplayName
+		}
+		appsRes = append(appsRes, &AppBean{Id: app.Id, Name: name})
+	}
+	return appsRes, err
+}
+
+func (impl PipelineBuilderImpl) GetAppListForEnvironment(request resourceGroup2.ResourceGroupingRequest) ([]*AppBean, error) {
+	var applicationList []*AppBean
+	var cdPipelines []*pipelineConfig.Pipeline
+	var err error
+	if request.ResourceGroupId > 0 {
+		appIds, err := impl.resourceGroupService.GetResourceIdsByResourceGroupId(request.ResourceGroupId)
+		if err != nil {
+			return nil, err
+		}
+		//override appIds if already provided app group id in request.
+		request.ResourceIds = appIds
+	}
+	if len(request.ResourceIds) > 0 {
+		cdPipelines, err = impl.pipelineRepository.FindActiveByInFilter(request.ParentResourceId, request.ResourceIds)
+	} else {
+		cdPipelines, err = impl.pipelineRepository.FindActiveByEnvId(request.ParentResourceId)
+	}
+	if err != nil {
+		impl.logger.Errorw("error in fetching pipelines", "request", request, "err", err)
+		return nil, err
+	}
+	if len(cdPipelines) == 0 {
+		return applicationList, nil
+	}
+	var appObjectArr []string
+	var envObjectArr []string
+	objects := impl.enforcerUtil.GetAppAndEnvObjectByDbPipeline(cdPipelines)
+	for _, object := range objects {
+		appObjectArr = append(appObjectArr, object[0])
+		envObjectArr = append(envObjectArr, object[1])
+	}
+	appResults, envResults := request.CheckAuthBatch(request.EmailId, appObjectArr, envObjectArr)
+	for _, pipeline := range cdPipelines {
+		appObject := objects[pipeline.Id][0]
+		envObject := objects[pipeline.Id][1]
+		if !(appResults[appObject] && envResults[envObject]) {
+			//if user unauthorized, skip items
+			continue
+		}
+		applicationList = append(applicationList, &AppBean{Id: pipeline.AppId, Name: pipeline.App.AppName})
+	}
+	return applicationList, err
+}
+
+func (impl *PipelineBuilderImpl) FindAppsByTeamId(teamId int) ([]*AppBean, error) {
+	var appsRes []*AppBean
+	apps, err := impl.appRepo.FindAppsByTeamId(teamId)
+	if err != nil {
+		impl.logger.Errorw("error while fetching app", "err", err)
+		return nil, err
+	}
+	for _, app := range apps {
+		appsRes = append(appsRes, &AppBean{Id: app.Id, Name: app.AppName})
+	}
+	return appsRes, err
+}
+
+func (impl *PipelineBuilderImpl) FindAppsByTeamName(teamName string) ([]AppBean, error) {
+	var appsRes []AppBean
+	apps, err := impl.appRepo.FindAppsByTeamName(teamName)
+	if err != nil {
+		impl.logger.Errorw("error while fetching app", "err", err)
+		return nil, err
+	}
+	for _, app := range apps {
+		appsRes = append(appsRes, AppBean{Id: app.Id, Name: app.AppName})
+	}
+	return appsRes, err
+}
 
 func (impl *PipelineBuilderImpl) getDefaultArtifactStore(id string) (store *dockerRegistryRepository.DockerArtifactStore, err error) {
 	if id == "" {
@@ -838,722 +696,6 @@ func (impl *PipelineBuilderImpl) getCiTemplateVariablesByAppIds(appIds []int) (m
 	return ciConfigMap, err
 }
 
-func (impl *PipelineBuilderImpl) GetTriggerViewCiPipeline(appId int) (*bean.TriggerViewCiConfig, error) {
-
-	triggerViewCiConfig := &bean.TriggerViewCiConfig{}
-
-	ciConfig, err := impl.getCiTemplateVariables(appId)
-	if err != nil {
-		impl.logger.Debugw("error in fetching ci pipeline", "appId", appId, "err", err)
-		return nil, err
-	}
-
-	triggerViewCiConfig.CiGitMaterialId = ciConfig.CiBuildConfig.GitMaterialId
-
-	// fetch pipelines
-	pipelines, err := impl.ciPipelineRepository.FindByAppId(appId)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching ci pipeline", "appId", appId, "err", err)
-		return nil, err
-	}
-
-	ciOverrideTemplateMap := make(map[int]*bean3.CiTemplateBean)
-	ciTemplateBeanOverrides, err := impl.ciTemplateService.FindTemplateOverrideByAppId(appId)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, templateBeanOverride := range ciTemplateBeanOverrides {
-		ciTemplateOverride := templateBeanOverride.CiTemplateOverride
-		ciOverrideTemplateMap[ciTemplateOverride.CiPipelineId] = templateBeanOverride
-	}
-
-	var ciPipelineResp []*bean.CiPipeline
-	for _, pipeline := range pipelines {
-		isLinkedCiPipeline := pipeline.IsExternal
-		ciPipeline := &bean.CiPipeline{
-			Id:                       pipeline.Id,
-			Version:                  pipeline.Version,
-			Name:                     pipeline.Name,
-			Active:                   pipeline.Active,
-			Deleted:                  pipeline.Deleted,
-			IsManual:                 pipeline.IsManual,
-			IsExternal:               isLinkedCiPipeline,
-			ParentCiPipeline:         pipeline.ParentCiPipeline,
-			ScanEnabled:              pipeline.ScanEnabled,
-			IsDockerConfigOverridden: pipeline.IsDockerConfigOverridden,
-			PipelineType:             bean.PipelineType(pipeline.PipelineType),
-		}
-		if ciTemplateBean, ok := ciOverrideTemplateMap[pipeline.Id]; ok {
-			templateOverride := ciTemplateBean.CiTemplateOverride
-			ciPipeline.DockerConfigOverride = bean.DockerConfigOverride{
-				DockerRegistry:   templateOverride.DockerRegistryId,
-				DockerRepository: templateOverride.DockerRepository,
-				CiBuildConfig:    ciTemplateBean.CiBuildConfig,
-			}
-		}
-		for _, material := range pipeline.CiPipelineMaterials {
-			// ignore those materials which have inactive git material
-			if material == nil || material.GitMaterial == nil || !material.GitMaterial.Active {
-				continue
-			}
-			ciMaterial := &bean.CiMaterial{
-				Id:              material.Id,
-				CheckoutPath:    material.CheckoutPath,
-				Path:            material.Path,
-				ScmId:           material.ScmId,
-				GitMaterialId:   material.GitMaterialId,
-				GitMaterialName: material.GitMaterial.Name[strings.Index(material.GitMaterial.Name, "-")+1:],
-				ScmName:         material.ScmName,
-				ScmVersion:      material.ScmVersion,
-				IsRegex:         material.Regex != "",
-				Source:          &bean.SourceTypeConfig{Type: material.Type, Value: material.Value, Regex: material.Regex},
-			}
-			ciPipeline.CiMaterial = append(ciPipeline.CiMaterial, ciMaterial)
-		}
-		linkedCis, err := impl.ciPipelineRepository.FindByParentCiPipelineId(ciPipeline.Id)
-		if err != nil && !util.IsErrNoRows(err) {
-			return nil, err
-		}
-		ciPipeline.LinkedCount = len(linkedCis)
-		ciPipelineResp = append(ciPipelineResp, ciPipeline)
-	}
-	triggerViewCiConfig.CiPipelines = ciPipelineResp
-	triggerViewCiConfig.Materials = ciConfig.Materials
-
-	return triggerViewCiConfig, nil
-}
-
-func (impl *PipelineBuilderImpl) GetCiPipeline(appId int) (ciConfig *bean.CiConfigRequest, err error) {
-	ciConfig, err = impl.getCiTemplateVariables(appId)
-	if err != nil {
-		impl.logger.Debugw("error in fetching ci pipeline", "appId", appId, "err", err)
-		return nil, err
-	}
-	//TODO fill these variables
-	//ciCdConfig.CiPipeline=
-	//--------pipeline population start
-	pipelines, err := impl.ciPipelineRepository.FindByAppId(appId)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching ci pipeline", "appId", appId, "err", err)
-		return nil, err
-	}
-
-	if impl.ciConfig.ExternalCiWebhookUrl == "" {
-		hostUrl, err := impl.attributesService.GetByKey(attributes.HostUrlKey)
-		if err != nil {
-			return nil, err
-		}
-		if hostUrl != nil {
-			impl.ciConfig.ExternalCiWebhookUrl = fmt.Sprintf("%s/%s", hostUrl.Value, ExternalCiWebhookPath)
-		}
-	}
-	//map of ciPipelineId and their templateOverrideConfig
-	ciOverrideTemplateMap := make(map[int]*bean3.CiTemplateBean)
-	ciTemplateBeanOverrides, err := impl.ciTemplateService.FindTemplateOverrideByAppId(appId)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, templateBeanOverride := range ciTemplateBeanOverrides {
-		ciTemplateOverride := templateBeanOverride.CiTemplateOverride
-		ciOverrideTemplateMap[ciTemplateOverride.CiPipelineId] = templateBeanOverride
-	}
-	var ciPipelineResp []*bean.CiPipeline
-	for _, pipeline := range pipelines {
-
-		dockerArgs := make(map[string]string)
-		if len(pipeline.DockerArgs) > 0 {
-			err := json.Unmarshal([]byte(pipeline.DockerArgs), &dockerArgs)
-			if err != nil {
-				impl.logger.Warnw("error in unmarshal", "err", err)
-			}
-		}
-
-		var externalCiConfig bean.ExternalCiConfig
-
-		ciPipelineScripts, err := impl.ciPipelineRepository.FindCiScriptsByCiPipelineId(pipeline.Id)
-		if err != nil && !util.IsErrNoRows(err) {
-			impl.logger.Errorw("error in fetching ci scripts")
-			return nil, err
-		}
-
-		var beforeDockerBuildScripts []*bean.CiScript
-		var afterDockerBuildScripts []*bean.CiScript
-		for _, ciScript := range ciPipelineScripts {
-			ciScriptResp := &bean.CiScript{
-				Id:             ciScript.Id,
-				Index:          ciScript.Index,
-				Name:           ciScript.Name,
-				Script:         ciScript.Script,
-				OutputLocation: ciScript.OutputLocation,
-			}
-			if ciScript.Stage == BEFORE_DOCKER_BUILD {
-				beforeDockerBuildScripts = append(beforeDockerBuildScripts, ciScriptResp)
-			} else if ciScript.Stage == AFTER_DOCKER_BUILD {
-				afterDockerBuildScripts = append(afterDockerBuildScripts, ciScriptResp)
-			}
-		}
-		parentCiPipeline, err := impl.ciPipelineRepository.FindById(pipeline.ParentCiPipeline)
-		if err != nil && !util.IsErrNoRows(err) {
-			impl.logger.Errorw("err", err)
-			return nil, err
-		}
-		ciPipeline := &bean.CiPipeline{
-			Id:                       pipeline.Id,
-			Version:                  pipeline.Version,
-			Name:                     pipeline.Name,
-			Active:                   pipeline.Active,
-			Deleted:                  pipeline.Deleted,
-			DockerArgs:               dockerArgs,
-			IsManual:                 pipeline.IsManual,
-			IsExternal:               pipeline.IsExternal,
-			ParentCiPipeline:         pipeline.ParentCiPipeline,
-			ParentAppId:              parentCiPipeline.AppId,
-			ExternalCiConfig:         externalCiConfig,
-			BeforeDockerBuildScripts: beforeDockerBuildScripts,
-			AfterDockerBuildScripts:  afterDockerBuildScripts,
-			ScanEnabled:              pipeline.ScanEnabled,
-			IsDockerConfigOverridden: pipeline.IsDockerConfigOverridden,
-			PipelineType:             bean.PipelineType(pipeline.PipelineType),
-		}
-		ciEnvMapping, err := impl.ciPipelineRepository.FindCiEnvMappingByCiPipelineId(pipeline.Id)
-		if err != nil && err != pg.ErrNoRows {
-			impl.logger.Errorw("error in fetching ciEnvMapping", "ciPipelineId ", pipeline.Id, "err", err)
-			return nil, err
-		}
-		if ciEnvMapping.Id > 0 {
-			ciPipeline.EnvironmentId = ciEnvMapping.EnvironmentId
-		}
-
-		lastTriggeredWorkflowEnv, err := impl.ciWorkflowRepository.FindLastTriggeredWorkflow(pipeline.Id)
-		if err != nil && err != pg.ErrNoRows {
-			impl.logger.Errorw("error in fetching lasTriggeredWorkflowEnv", "ciPipelineId ", pipeline.Id, "err", err)
-			return nil, err
-		}
-		if err == pg.ErrNoRows {
-			ciPipeline.LastTriggeredEnvId = -1
-		} else {
-			ciPipeline.LastTriggeredEnvId = lastTriggeredWorkflowEnv.EnvironmentId
-		}
-
-		if ciTemplateBean, ok := ciOverrideTemplateMap[pipeline.Id]; ok {
-			templateOverride := ciTemplateBean.CiTemplateOverride
-			ciPipeline.DockerConfigOverride = bean.DockerConfigOverride{
-				DockerRegistry:   templateOverride.DockerRegistryId,
-				DockerRepository: templateOverride.DockerRepository,
-				CiBuildConfig:    ciTemplateBean.CiBuildConfig,
-			}
-		}
-		for _, material := range pipeline.CiPipelineMaterials {
-			// ignore those materials which have inactive git material
-			if material == nil || material.GitMaterial == nil || !material.GitMaterial.Active {
-				continue
-			}
-			ciMaterial := &bean.CiMaterial{
-				Id:              material.Id,
-				CheckoutPath:    material.CheckoutPath,
-				Path:            material.Path,
-				ScmId:           material.ScmId,
-				GitMaterialId:   material.GitMaterialId,
-				GitMaterialName: material.GitMaterial.Name[strings.Index(material.GitMaterial.Name, "-")+1:],
-				ScmName:         material.ScmName,
-				ScmVersion:      material.ScmVersion,
-				IsRegex:         material.Regex != "",
-				Source:          &bean.SourceTypeConfig{Type: material.Type, Value: material.Value, Regex: material.Regex},
-			}
-			ciPipeline.CiMaterial = append(ciPipeline.CiMaterial, ciMaterial)
-		}
-		linkedCis, err := impl.ciPipelineRepository.FindByParentCiPipelineId(ciPipeline.Id)
-		if err != nil && !util.IsErrNoRows(err) {
-			return nil, err
-		}
-		ciPipeline.LinkedCount = len(linkedCis)
-		ciPipelineResp = append(ciPipelineResp, ciPipeline)
-	}
-	ciConfig.CiPipelines = ciPipelineResp
-	//--------pipeline population end
-	return ciConfig, err
-}
-
-func (impl *PipelineBuilderImpl) GetExternalCi(appId int) (ciConfig []*bean.ExternalCiConfig, err error) {
-	externalCiPipelines, err := impl.ciPipelineRepository.FindExternalCiByAppId(appId)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching external ci", "appId", appId, "err", err)
-		return nil, err
-	}
-
-	hostUrl, err := impl.attributesService.GetByKey(attributes.HostUrlKey)
-	if err != nil {
-		impl.logger.Errorw("error in fetching external ci", "appId", appId, "err", err)
-		return nil, err
-	}
-	if hostUrl != nil {
-		impl.ciConfig.ExternalCiWebhookUrl = fmt.Sprintf("%s/%s", hostUrl.Value, ExternalCiWebhookPath)
-	}
-
-	externalCiConfigs := make([]*bean.ExternalCiConfig, 0)
-
-	var externalCiPipelineIds []int
-	appWorkflowMappingsMap := make(map[int][]*appWorkflow.AppWorkflowMapping)
-
-	for _, externalCiPipeline := range externalCiPipelines {
-		externalCiPipelineIds = append(externalCiPipelineIds, externalCiPipeline.Id)
-	}
-	if len(externalCiPipelineIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no external ci pipeline found"}
-		return externalCiConfigs, err
-	}
-	appWorkflowMappings, err := impl.appWorkflowRepository.FindWFCDMappingByExternalCiIdByIdsIn(externalCiPipelineIds)
-	if err != nil {
-		impl.logger.Errorw("Error in fetching app workflow mapping for CD pipeline by external CI ID", "err", err)
-		return nil, err
-	}
-
-	for _, appWorkflowMapping := range appWorkflowMappings {
-		appWorkflowMappingsMap[appWorkflowMapping.ParentId] = append(appWorkflowMappingsMap[appWorkflowMapping.ParentId], appWorkflowMapping)
-	}
-
-	for _, externalCiPipeline := range externalCiPipelines {
-		externalCiConfig := &bean.ExternalCiConfig{
-			Id:         externalCiPipeline.Id,
-			WebhookUrl: fmt.Sprintf("%s/%d", impl.ciConfig.ExternalCiWebhookUrl, externalCiPipeline.Id),
-			Payload:    impl.ciConfig.ExternalCiPayload,
-			AccessKey:  "",
-		}
-
-		if _, ok := appWorkflowMappingsMap[externalCiPipeline.Id]; !ok {
-			impl.logger.Errorw("unable to find app workflow cd mapping corresponding to external ci pipeline id")
-			return nil, errors.New("unable to find app workflow cd mapping corresponding to external ci pipeline id")
-		}
-
-		var appWorkflowComponentIds []int
-		var appIds []int
-
-		CDPipelineMap := make(map[int]*pipelineConfig.Pipeline)
-		appIdMap := make(map[int]*app2.App)
-
-		for _, appWorkflowMappings := range appWorkflowMappings {
-			appWorkflowComponentIds = append(appWorkflowComponentIds, appWorkflowMappings.ComponentId)
-		}
-		if len(appWorkflowComponentIds) == 0 {
-			continue
-		}
-		cdPipelines, err := impl.pipelineRepository.FindAppAndEnvironmentAndProjectByPipelineIds(appWorkflowComponentIds)
-		if err != nil && !util.IsErrNoRows(err) {
-			impl.logger.Errorw("error in fetching external ci", "appId", appId, "err", err)
-			return nil, err
-		}
-		for _, pipeline := range cdPipelines {
-			CDPipelineMap[pipeline.Id] = pipeline
-			appIds = append(appIds, pipeline.AppId)
-		}
-		if len(appIds) == 0 {
-			continue
-		}
-		apps, err := impl.appRepo.FindAppAndProjectByIdsIn(appIds)
-		for _, app := range apps {
-			appIdMap[app.Id] = app
-		}
-
-		roleData := make(map[string]interface{})
-		for _, appWorkflowMapping := range appWorkflowMappings {
-			if _, ok := CDPipelineMap[appWorkflowMapping.ComponentId]; !ok {
-				impl.logger.Errorw("error in getting cd pipeline data for workflow", "app workflow id", appWorkflowMapping.ComponentId, "err", err)
-				return nil, errors.New("error in getting cd pipeline data for workflow")
-			}
-			cdPipeline := CDPipelineMap[appWorkflowMapping.ComponentId]
-
-			if _, ok := roleData[teamIdKey]; !ok {
-				if _, ok := appIdMap[cdPipeline.AppId]; !ok {
-					impl.logger.Errorw("error in getting app data for pipeline", "app id", cdPipeline.AppId)
-					return nil, errors.New("error in getting app data for pipeline")
-				}
-				app := appIdMap[cdPipeline.AppId]
-				roleData[teamIdKey] = app.TeamId
-				roleData[teamNameKey] = app.Team.Name
-				roleData[appIdKey] = cdPipeline.AppId
-				roleData[appNameKey] = cdPipeline.App.AppName
-			}
-			if _, ok := roleData[environmentNameKey]; !ok {
-				roleData[environmentNameKey] = cdPipeline.Environment.Name
-			} else {
-				roleData[environmentNameKey] = fmt.Sprintf("%s,%s", roleData[environmentNameKey], cdPipeline.Environment.Name)
-			}
-			if _, ok := roleData[environmentIdentifierKey]; !ok {
-				roleData[environmentIdentifierKey] = cdPipeline.Environment.EnvironmentIdentifier
-			} else {
-				roleData[environmentIdentifierKey] = fmt.Sprintf("%s,%s", roleData[environmentIdentifierKey], cdPipeline.Environment.EnvironmentIdentifier)
-			}
-		}
-
-		externalCiConfig.ExternalCiConfigRole = bean.ExternalCiConfigRole{
-			ProjectId:             roleData[teamIdKey].(int),
-			ProjectName:           roleData[teamNameKey].(string),
-			AppId:                 roleData[appIdKey].(int),
-			AppName:               roleData[appNameKey].(string),
-			EnvironmentName:       roleData[environmentNameKey].(string),
-			EnvironmentIdentifier: roleData[environmentIdentifierKey].(string),
-			Role:                  "Build and deploy",
-		}
-		externalCiConfigs = append(externalCiConfigs, externalCiConfig)
-	}
-	//--------pipeline population end
-	return externalCiConfigs, err
-}
-
-func (impl *PipelineBuilderImpl) GetExternalCiById(appId int, externalCiId int) (ciConfig *bean.ExternalCiConfig, err error) {
-
-	externalCiPipeline, err := impl.ciPipelineRepository.FindExternalCiById(externalCiId)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching external ci", "appId", appId, "err", err)
-		return nil, err
-	}
-
-	if externalCiPipeline.Id == 0 {
-		impl.logger.Errorw("invalid external ci id", "externalCiId", externalCiId, "err", err)
-		return nil, &util.ApiError{Code: "400", HttpStatusCode: 400, UserMessage: "invalid external ci id"}
-	}
-
-	hostUrl, err := impl.attributesService.GetByKey(attributes.HostUrlKey)
-	if err != nil {
-		impl.logger.Errorw("error in fetching external ci", "appId", appId, "err", err)
-		return nil, err
-	}
-	if hostUrl != nil {
-		impl.ciConfig.ExternalCiWebhookUrl = fmt.Sprintf("%s/%s", hostUrl.Value, ExternalCiWebhookPath)
-	}
-
-	appWorkflowMappings, err := impl.appWorkflowRepository.FindWFCDMappingByExternalCiId(externalCiPipeline.Id)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching external ci", "appId", appId, "err", err)
-		return nil, err
-	}
-
-	roleData := make(map[string]interface{})
-	for _, appWorkflowMapping := range appWorkflowMappings {
-		cdPipeline, err := impl.pipelineRepository.FindById(appWorkflowMapping.ComponentId)
-		if err != nil && !util.IsErrNoRows(err) {
-			impl.logger.Errorw("error in fetching external ci", "appId", appId, "err", err)
-			return nil, err
-		}
-		if _, ok := roleData[teamIdKey]; !ok {
-			app, err := impl.appRepo.FindAppAndProjectByAppId(cdPipeline.AppId)
-			if err != nil && !util.IsErrNoRows(err) {
-				impl.logger.Errorw("error in fetching external ci", "appId", appId, "err", err)
-				return nil, err
-			}
-			roleData[teamIdKey] = app.TeamId
-			roleData[teamNameKey] = app.Team.Name
-			roleData[appIdKey] = cdPipeline.AppId
-			roleData[appNameKey] = cdPipeline.App.AppName
-		}
-		if _, ok := roleData[environmentNameKey]; !ok {
-			roleData[environmentNameKey] = cdPipeline.Environment.Name
-		} else {
-			roleData[environmentNameKey] = fmt.Sprintf("%s,%s", roleData[environmentNameKey], cdPipeline.Environment.Name)
-		}
-		if _, ok := roleData[environmentIdentifierKey]; !ok {
-			roleData[environmentIdentifierKey] = cdPipeline.Environment.EnvironmentIdentifier
-		} else {
-			roleData[environmentIdentifierKey] = fmt.Sprintf("%s,%s", roleData[environmentIdentifierKey], cdPipeline.Environment.EnvironmentIdentifier)
-		}
-	}
-
-	externalCiConfig := &bean.ExternalCiConfig{
-		Id:         externalCiPipeline.Id,
-		WebhookUrl: fmt.Sprintf("%s/%d", impl.ciConfig.ExternalCiWebhookUrl, externalCiId),
-		Payload:    impl.ciConfig.ExternalCiPayload,
-		AccessKey:  "",
-	}
-	externalCiConfig.ExternalCiConfigRole = bean.ExternalCiConfigRole{
-		ProjectId:             roleData[teamIdKey].(int),
-		ProjectName:           roleData[teamNameKey].(string),
-		AppId:                 roleData[appIdKey].(int),
-		AppName:               roleData[appNameKey].(string),
-		EnvironmentName:       roleData[environmentNameKey].(string),
-		EnvironmentIdentifier: roleData[environmentIdentifierKey].(string),
-		Role:                  "Build and deploy",
-	}
-	externalCiConfig.Schema = impl.buildExternalCiWebhookSchema()
-	externalCiConfig.PayloadOption = impl.buildPayloadOption()
-	externalCiConfig.Responses = impl.buildResponses()
-	//--------pipeline population end
-	return externalCiConfig, err
-}
-
-func (impl *PipelineBuilderImpl) GetCiPipelineMin(appId int, envIds []int) ([]*bean.CiPipelineMin, error) {
-	pipelines := make([]*pipelineConfig.CiPipeline, 0)
-	var err error
-	if len(envIds) > 0 {
-		//filter ci_pipelines based on env list
-		pipelines, err = impl.ciPipelineRepository.FindCiPipelineByAppIdAndEnvIds(appId, envIds)
-	} else {
-		pipelines, err = impl.ciPipelineRepository.FindByAppId(appId)
-	}
-
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("error in fetching ci pipeline", "appId", appId, "err", err)
-		return nil, err
-	}
-	if err == pg.ErrNoRows || len(pipelines) == 0 {
-		impl.logger.Errorw("no ci pipeline found", "appId", appId, "err", err)
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no ci pipeline found"}
-		return nil, err
-	}
-	parentCiPipelines, linkedCiPipelineIds, err := impl.ciPipelineRepository.FindParentCiPipelineMapByAppId(appId)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("err", err)
-		return nil, err
-	}
-	pipelineParentCiMap := make(map[int]*pipelineConfig.CiPipeline)
-	for index, item := range parentCiPipelines {
-		pipelineParentCiMap[linkedCiPipelineIds[index]] = item
-	}
-
-	var ciPipelineResp []*bean.CiPipelineMin
-	for _, pipeline := range pipelines {
-		parentCiPipeline := pipelineConfig.CiPipeline{}
-		pipelineType := bean.NORMAL
-
-		if pipelineParentCiMap[pipeline.Id] != nil {
-			parentCiPipeline = *pipelineParentCiMap[pipeline.Id]
-			pipelineType = bean.LINKED
-		} else if pipeline.IsExternal == true {
-			pipelineType = bean.EXTERNAL
-		} else if pipeline.PipelineType == string(bean.CI_JOB) {
-			pipelineType = bean.CI_JOB
-		}
-
-		ciPipeline := &bean.CiPipelineMin{
-			Id:               pipeline.Id,
-			Name:             pipeline.Name,
-			ParentCiPipeline: pipeline.ParentCiPipeline,
-			ParentAppId:      parentCiPipeline.AppId,
-			PipelineType:     pipelineType,
-			ScanEnabled:      pipeline.ScanEnabled,
-		}
-		ciPipelineResp = append(ciPipelineResp, ciPipeline)
-	}
-	return ciPipelineResp, err
-}
-
-func (impl PipelineBuilderImpl) UpdateCiTemplate(updateRequest *bean.CiConfigRequest) (*bean.CiConfigRequest, error) {
-	originalCiConf, err := impl.getCiTemplateVariables(updateRequest.AppId)
-	if err != nil {
-		impl.logger.Errorw("error in fetching original ciCdConfig for update", "appId", updateRequest.Id, "err", err)
-		return nil, err
-	}
-	if originalCiConf.Version != updateRequest.Version {
-		impl.logger.Errorw("stale version requested", "appId", updateRequest.Id, "old", originalCiConf.Version, "new", updateRequest.Version)
-		return nil, fmt.Errorf("stale version of resource requested kindly refresh. requested: %s, found %s", updateRequest.Version, originalCiConf.Version)
-	}
-	dockerArtifaceStore, err := impl.dockerArtifactStoreRepository.FindOne(updateRequest.DockerRegistry)
-	if err != nil {
-		impl.logger.Errorw("error in fetching DockerRegistry  for update", "appId", updateRequest.Id, "err", err, "registry", updateRequest.DockerRegistry)
-		return nil, err
-	}
-	regHost, err := dockerArtifaceStore.GetRegistryLocation()
-	if err != nil {
-		impl.logger.Errorw("invalid reg url", "err", err)
-		return nil, err
-	}
-
-	var repo string
-	if updateRequest.DockerRepository != "" {
-		repo = updateRequest.DockerRepository
-	} else {
-		repo = originalCiConf.DockerRepository
-	}
-
-	if dockerArtifaceStore.RegistryType == dockerRegistryRepository.REGISTRYTYPE_ECR {
-		err := impl.ciCdPipelineOrchestrator.CreateEcrRepo(repo, dockerArtifaceStore.AWSRegion, dockerArtifaceStore.AWSAccessKeyId, dockerArtifaceStore.AWSSecretAccessKey)
-		if err != nil {
-			impl.logger.Errorw("ecr repo creation failed while updating ci template", "repo", repo, "err", err)
-			return nil, err
-		}
-	}
-
-	originalCiConf.AfterDockerBuild = updateRequest.AfterDockerBuild
-	originalCiConf.BeforeDockerBuild = updateRequest.BeforeDockerBuild
-	//originalCiConf.CiBuildConfigBean = updateRequest.CiBuildConfigBean
-	originalCiConf.DockerRegistry = updateRequest.DockerRegistry
-	originalCiConf.DockerRepository = updateRequest.DockerRepository
-	originalCiConf.DockerRegistryUrl = regHost
-
-	//argByte, err := json.Marshal(originalCiConf.DockerBuildConfig.Args)
-	//if err != nil {
-	//	return nil, err
-	//}
-	afterByte, err := json.Marshal(originalCiConf.AfterDockerBuild)
-	if err != nil {
-		return nil, err
-	}
-	beforeByte, err := json.Marshal(originalCiConf.BeforeDockerBuild)
-	if err != nil {
-		return nil, err
-	}
-	//buildOptionsByte, err := json.Marshal(originalCiConf.DockerBuildConfig.DockerBuildOptions)
-	//if err != nil {
-	//	impl.logger.Errorw("error in marshaling dockerBuildOptions", "err", err)
-	//	return nil, err
-	//}
-	ciBuildConfig := updateRequest.CiBuildConfig
-	originalCiBuildConfig := originalCiConf.CiBuildConfig
-	ciTemplate := &pipelineConfig.CiTemplate{
-		//DockerfilePath:    originalCiConf.DockerBuildConfig.DockerfilePath,
-		GitMaterialId:             ciBuildConfig.GitMaterialId,
-		BuildContextGitMaterialId: ciBuildConfig.BuildContextGitMaterialId,
-		//Args:              string(argByte),
-		//TargetPlatform:    originalCiConf.DockerBuildConfig.TargetPlatform,
-		AppId:             originalCiConf.AppId,
-		BeforeDockerBuild: string(beforeByte),
-		AfterDockerBuild:  string(afterByte),
-		Version:           originalCiConf.Version,
-		Id:                originalCiConf.Id,
-		DockerRepository:  originalCiConf.DockerRepository,
-		DockerRegistryId:  &originalCiConf.DockerRegistry,
-		Active:            true,
-		AuditLog: sql.AuditLog{
-			CreatedOn: originalCiConf.CreatedOn,
-			CreatedBy: originalCiConf.CreatedBy,
-			UpdatedOn: time.Now(),
-			UpdatedBy: updateRequest.UserId,
-		},
-	}
-
-	ciBuildConfig.Id = originalCiBuildConfig.Id
-	ciTemplateBean := &bean3.CiTemplateBean{
-		CiTemplate:    ciTemplate,
-		CiBuildConfig: ciBuildConfig,
-		UserId:        updateRequest.UserId,
-	}
-	err = impl.ciTemplateService.Update(ciTemplateBean)
-	if err != nil {
-		return nil, err
-	}
-
-	originalCiConf.CiBuildConfig = ciBuildConfig
-
-	err = impl.CiTemplateHistoryService.SaveHistory(ciTemplateBean, "update")
-
-	if err != nil {
-		impl.logger.Errorw("error in saving update history for ci template", "error", err)
-	}
-
-	return originalCiConf, nil
-}
-
-func (impl PipelineBuilderImpl) CreateCiPipeline(createRequest *bean.CiConfigRequest) (*bean.PipelineCreateResponse, error) {
-	impl.logger.Debugw("pipeline create request received", "req", createRequest)
-
-	//-----------fetch data
-	app, err := impl.appRepo.FindById(createRequest.AppId)
-	if err != nil {
-		impl.logger.Errorw("error in fetching pipeline group", "groupId", createRequest.AppId, "err", err)
-		return nil, err
-	}
-	//--ecr config
-	createRequest.AppName = app.AppName
-	if !createRequest.IsJob {
-		store, err := impl.getDefaultArtifactStore(createRequest.DockerRegistry)
-		if err != nil {
-			impl.logger.Errorw("error in fetching docker store ", "id", createRequest.DockerRepository, "err", err)
-			return nil, err
-		}
-
-		regHost, err := store.GetRegistryLocation()
-		if err != nil {
-			impl.logger.Errorw("invalid reg url", "err", err)
-			return nil, err
-		}
-		createRequest.DockerRegistryUrl = regHost
-		createRequest.DockerRegistry = store.Id
-
-		var repo string
-		if createRequest.DockerRepository != "" {
-			repo = createRequest.DockerRepository
-		} else {
-			repo = impl.ecrConfig.EcrPrefix + app.AppName
-		}
-
-		if store.RegistryType == dockerRegistryRepository.REGISTRYTYPE_ECR {
-			err := impl.ciCdPipelineOrchestrator.CreateEcrRepo(repo, store.AWSRegion, store.AWSAccessKeyId, store.AWSSecretAccessKey)
-			if err != nil {
-				impl.logger.Errorw("ecr repo creation failed while creating ci pipeline", "repo", repo, "err", err)
-				return nil, err
-			}
-		}
-		createRequest.DockerRepository = repo
-	}
-	//--ecr config	end
-	//-- template config start
-
-	//argByte, err := json.Marshal(createRequest.DockerBuildConfig.Args)
-	//if err != nil {
-	//	return nil, err
-	//}
-	afterByte, err := json.Marshal(createRequest.AfterDockerBuild)
-	if err != nil {
-		return nil, err
-	}
-	beforeByte, err := json.Marshal(createRequest.BeforeDockerBuild)
-	if err != nil {
-		return nil, err
-	}
-	buildConfig := createRequest.CiBuildConfig
-	ciTemplate := &pipelineConfig.CiTemplate{
-		//DockerRegistryId: createRequest.DockerRegistry,
-		//DockerRepository: createRequest.DockerRepository,
-		GitMaterialId:             buildConfig.GitMaterialId,
-		BuildContextGitMaterialId: buildConfig.BuildContextGitMaterialId,
-		//DockerfilePath:    createRequest.DockerBuildConfig.DockerfilePath,
-		//Args:              string(argByte),
-		//TargetPlatform:    createRequest.DockerBuildConfig.TargetPlatform,
-		Active:            true,
-		TemplateName:      createRequest.CiTemplateName,
-		Version:           createRequest.Version,
-		AppId:             createRequest.AppId,
-		AfterDockerBuild:  string(afterByte),
-		BeforeDockerBuild: string(beforeByte),
-		AuditLog:          sql.AuditLog{CreatedOn: time.Now(), UpdatedOn: time.Now(), CreatedBy: createRequest.UserId, UpdatedBy: createRequest.UserId},
-	}
-	if !createRequest.IsJob {
-		ciTemplate.DockerRegistryId = &createRequest.DockerRegistry
-		ciTemplate.DockerRepository = createRequest.DockerRepository
-	}
-
-	ciTemplateBean := &bean3.CiTemplateBean{
-		CiTemplate:    ciTemplate,
-		CiBuildConfig: createRequest.CiBuildConfig,
-	}
-	err = impl.ciTemplateService.Save(ciTemplateBean)
-	if err != nil {
-		return nil, err
-	}
-
-	//-- template config end
-
-	err = impl.CiTemplateHistoryService.SaveHistory(ciTemplateBean, "add")
-
-	if err != nil {
-		impl.logger.Errorw("error in saving audit logs of ci Template", "error", err)
-	}
-
-	createRequest.Id = ciTemplate.Id
-	createRequest.CiTemplateName = ciTemplate.TemplateName
-	if len(createRequest.CiPipelines) > 0 {
-		conf, err := impl.addpipelineToTemplate(createRequest)
-		if err != nil {
-			impl.logger.Errorw("error in pipeline creation ", "err", err)
-			return nil, err
-		}
-		impl.logger.Debugw("pipeline created ", "detail", conf)
-	}
-	createRes := &bean.PipelineCreateResponse{AppName: app.AppName, AppId: createRequest.AppId} //FIXME
-	return createRes, nil
-}
-
 func (impl *PipelineBuilderImpl) getGitMaterialsForApp(appId int) ([]*bean.GitMaterial, error) {
 	materials, err := impl.materialRepo.FindByAppId(appId)
 	if err != nil {
@@ -1651,212 +793,6 @@ func (impl *PipelineBuilderImpl) addpipelineToTemplate(createRequest *bean.CiCon
 	return createRequest, err
 }
 
-func (impl *PipelineBuilderImpl) PatchCiPipeline(request *bean.CiPatchRequest) (ciConfig *bean.CiConfigRequest, err error) {
-	ciConfig, err = impl.getCiTemplateVariables(request.AppId)
-	if err != nil {
-		impl.logger.Errorw("err in fetching template for pipeline patch, ", "err", err, "appId", request.AppId)
-		return nil, err
-	}
-	ciConfig.AppWorkflowId = request.AppWorkflowId
-	ciConfig.UserId = request.UserId
-	if request.CiPipeline != nil {
-		//setting ScanEnabled value from env variable,
-		request.CiPipeline.ScanEnabled = request.CiPipeline.ScanEnabled || impl.securityConfig.ForceSecurityScanning
-		ciConfig.ScanEnabled = request.CiPipeline.ScanEnabled
-	}
-
-	ciConfig.IsJob = request.IsJob
-	// Check for clone job to not create env override again
-	ciConfig.IsCloneJob = request.IsCloneJob
-
-	switch request.Action {
-	case bean.CREATE:
-		impl.logger.Debugw("create patch request")
-		ciConfig.CiPipelines = []*bean.CiPipeline{request.CiPipeline} //request.CiPipeline
-
-		res, err := impl.addpipelineToTemplate(ciConfig)
-		if err != nil {
-			impl.logger.Errorw("error in adding pipeline to template", "ciConf", ciConfig, "err", err)
-			return nil, err
-		}
-		return res, nil
-	case bean.UPDATE_SOURCE:
-		return impl.patchCiPipelineUpdateSource(ciConfig, request.CiPipeline)
-	case bean.DELETE:
-		pipeline, err := impl.DeleteCiPipeline(request)
-		if err != nil {
-			return nil, err
-		}
-		ciConfig.CiPipelines = []*bean.CiPipeline{pipeline}
-		return ciConfig, nil
-	default:
-		impl.logger.Errorw("unsupported operation ", "op", request.Action)
-		return nil, fmt.Errorf("unsupported operation %s", request.Action)
-	}
-
-}
-
-func (impl *PipelineBuilderImpl) PatchRegexCiPipeline(request *bean.CiRegexPatchRequest) (err error) {
-	var materials []*pipelineConfig.CiPipelineMaterial
-	for _, material := range request.CiPipelineMaterial {
-		materialDbObject, err := impl.ciPipelineMaterialRepository.GetById(material.Id)
-		if err != nil {
-			impl.logger.Errorw("err in fetching material, ", "err", err)
-			return err
-		}
-		if materialDbObject.Regex != "" {
-			if !impl.ciCdPipelineOrchestrator.CheckStringMatchRegex(materialDbObject.Regex, material.Value) {
-				impl.logger.Errorw("not matching given regex, ", "err", err)
-				return errors.New("not matching given regex")
-			}
-		}
-		pipelineMaterial := &pipelineConfig.CiPipelineMaterial{
-			Id:            material.Id,
-			Value:         material.Value,
-			CiPipelineId:  materialDbObject.CiPipelineId,
-			Type:          pipelineConfig.SourceType(material.Type),
-			Active:        true,
-			GitMaterialId: materialDbObject.GitMaterialId,
-			Regex:         materialDbObject.Regex,
-			AuditLog:      sql.AuditLog{UpdatedBy: request.UserId, UpdatedOn: time.Now(), CreatedOn: time.Now(), CreatedBy: request.UserId},
-		}
-		materials = append(materials, pipelineMaterial)
-	}
-
-	dbConnection := impl.pipelineRepository.GetConnection()
-	tx, err := dbConnection.Begin()
-	if err != nil {
-		return err
-	}
-	// Rollback tx on error.
-	defer tx.Rollback()
-
-	err = impl.ciPipelineMaterialRepository.Update(tx, materials...)
-	if err != nil {
-		return err
-	}
-
-	err = tx.Commit()
-	if err != nil {
-		return err
-	}
-
-	err = impl.ciCdPipelineOrchestrator.AddPipelineMaterialInGitSensor(materials)
-	if err != nil {
-		impl.logger.Errorf("error in saving pipelineMaterials in git sensor", "materials", materials, "err", err)
-		return err
-	}
-	return nil
-}
-func (impl *PipelineBuilderImpl) DeleteCiPipeline(request *bean.CiPatchRequest) (*bean.CiPipeline, error) {
-	ciPipelineId := request.CiPipeline.Id
-	//wf validation
-	workflowMapping, err := impl.appWorkflowRepository.FindWFCDMappingByCIPipelineId(ciPipelineId)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("error in fetching workflow mapping for ci validation", "err", err)
-		return nil, err
-	}
-	if len(workflowMapping) > 0 {
-		return nil, &util.ApiError{
-			InternalMessage:   "cd pipeline exists for this CI",
-			UserDetailMessage: fmt.Sprintf("cd pipeline exists for this CI"),
-			UserMessage:       fmt.Sprintf("cd pipeline exists for this CI")}
-	}
-
-	pipeline, err := impl.ciPipelineRepository.FindById(ciPipelineId)
-	if err != nil {
-		impl.logger.Errorw("pipeline fetch err", "id", ciPipelineId, "err", err)
-		return nil, err
-	}
-	appId := request.AppId
-	if pipeline.AppId != appId {
-		return nil, fmt.Errorf("invalid appid: %d pipelineId: %d mapping", appId, ciPipelineId)
-	}
-
-	dbConnection := impl.pipelineRepository.GetConnection()
-	tx, err := dbConnection.Begin()
-	if err != nil {
-		return nil, err
-	}
-	// Rollback tx on error.
-	defer tx.Rollback()
-
-	err = impl.ciCdPipelineOrchestrator.DeleteCiPipeline(pipeline, request, tx)
-	if err != nil {
-		impl.logger.Errorw("error in deleting pipeline db")
-		return nil, err
-	}
-
-	//delete app workflow mapping
-	appWorkflowMappings, err := impl.appWorkflowRepository.FindWFCIMappingByCIPipelineId(pipeline.Id)
-	for _, mapping := range appWorkflowMappings {
-		err := impl.appWorkflowRepository.DeleteAppWorkflowMapping(mapping, tx)
-		if err != nil {
-			impl.logger.Errorw("error in deleting workflow mapping", "err", err)
-			return nil, err
-		}
-	}
-	if request.CiPipeline.PreBuildStage != nil && request.CiPipeline.PreBuildStage.Id > 0 {
-		//deleting pre stage
-		err = impl.pipelineStageService.DeletePipelineStage(request.CiPipeline.PreBuildStage, request.UserId, tx)
-		if err != nil {
-			impl.logger.Errorw("error in deleting pre stage", "err", err, "preBuildStage", request.CiPipeline.PreBuildStage)
-			return nil, err
-		}
-	}
-	if request.CiPipeline.PostBuildStage != nil && request.CiPipeline.PostBuildStage.Id > 0 {
-		//deleting post stage
-		err = impl.pipelineStageService.DeletePipelineStage(request.CiPipeline.PostBuildStage, request.UserId, tx)
-		if err != nil {
-			impl.logger.Errorw("error in deleting post stage", "err", err, "postBuildStage", request.CiPipeline.PostBuildStage)
-			return nil, err
-		}
-	}
-	err = tx.Commit()
-	if err != nil {
-		return nil, err
-	}
-	request.CiPipeline.Deleted = true
-	request.CiPipeline.Name = pipeline.Name
-	return request.CiPipeline, nil
-	//delete pipeline
-	//delete scm
-
-}
-
-func (impl *PipelineBuilderImpl) PatchCiMaterialSource(ciPipeline *bean.CiMaterialPatchRequest, userId int32) (*bean.CiMaterialPatchRequest, error) {
-	return impl.ciCdPipelineOrchestrator.PatchCiMaterialSource(ciPipeline, userId)
-}
-
-func (impl *PipelineBuilderImpl) BulkPatchCiMaterialSource(ciPipelines *bean.CiMaterialBulkPatchRequest, userId int32, token string, checkAppSpecificAccess func(token, action string, appId int) (bool, error)) (*bean.CiMaterialBulkPatchResponse, error) {
-	response := &bean.CiMaterialBulkPatchResponse{}
-	var ciPipelineMaterials []*pipelineConfig.CiPipelineMaterial
-	for _, appId := range ciPipelines.AppIds {
-		ciPipeline := &bean.CiMaterialValuePatchRequest{
-			AppId:         appId,
-			EnvironmentId: ciPipelines.EnvironmentId,
-		}
-		ciPipelineMaterial, err := impl.ciCdPipelineOrchestrator.PatchCiMaterialSourceValue(ciPipeline, userId, ciPipelines.Value, token, checkAppSpecificAccess)
-
-		if err == nil {
-			ciPipelineMaterial.Type = pipelineConfig.SOURCE_TYPE_BRANCH_FIXED
-			ciPipelineMaterials = append(ciPipelineMaterials, ciPipelineMaterial)
-		}
-		response.Apps = append(response.Apps, bean.CiMaterialPatchResponse{
-			AppId:   appId,
-			Status:  getPatchStatus(err),
-			Message: getPatchMessage(err),
-		})
-	}
-	if len(ciPipelineMaterials) == 0 {
-		return response, nil
-	}
-	if err := impl.ciCdPipelineOrchestrator.UpdateCiPipelineMaterials(ciPipelineMaterials); err != nil {
-		return nil, err
-	}
-	return response, nil
-}
-
 func getPatchStatus(err error) bean.CiPatchStatus {
 	if err != nil {
 		if err.Error() == string(bean.CI_PATCH_NOT_AUTHORIZED_MESSAGE) {
@@ -1902,23 +838,6 @@ func (impl *PipelineBuilderImpl) patchCiPipelineUpdateSource(baseCiConfig *bean.
 		baseCiConfig.CiPipelines = append(baseCiConfig.CiPipelines, modifiedCiPipeline)
 		return baseCiConfig, err
 	}
-
-}
-
-func (impl *PipelineBuilderImpl) IsGitopsConfigured() (bool, error) {
-
-	isGitOpsConfigured := false
-	gitOpsConfig, err := impl.gitOpsRepository.GetGitOpsConfigActive()
-
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("GetGitOpsConfigActive, error while getting", "err", err)
-		return false, err
-	}
-	if gitOpsConfig != nil && gitOpsConfig.Id > 0 {
-		isGitOpsConfigured = true
-	}
-
-	return isGitOpsConfigured, nil
 
 }
 
@@ -2009,128 +928,6 @@ func (impl *PipelineBuilderImpl) RegisterInACD(gitOpsRepoName string, chartGitAt
 	return nil
 }
 
-func (impl *PipelineBuilderImpl) IsGitOpsRequiredForCD(pipelineCreateRequest *bean.CdPipelines) bool {
-
-	// if deploymentAppType is not coming in request than hasAtLeastOneGitOps will be false
-
-	haveAtLeastOneGitOps := false
-	for _, pipeline := range pipelineCreateRequest.Pipelines {
-		if pipeline.DeploymentAppType == util.PIPELINE_DEPLOYMENT_TYPE_ACD {
-			haveAtLeastOneGitOps = true
-		}
-	}
-	return haveAtLeastOneGitOps
-}
-
-func (impl *PipelineBuilderImpl) SetPipelineDeploymentAppType(pipelineCreateRequest *bean.CdPipelines, isGitOpsConfigured bool, deploymentTypeValidationConfig map[string]bool) {
-	for _, pipeline := range pipelineCreateRequest.Pipelines {
-		// by default both deployment app type are allowed
-		AllowedDeploymentAppTypes := map[string]bool{
-			util.PIPELINE_DEPLOYMENT_TYPE_ACD:  true,
-			util.PIPELINE_DEPLOYMENT_TYPE_HELM: true,
-		}
-		for k, v := range deploymentTypeValidationConfig {
-			// rewriting allowed deployment types based on config provided by user
-			AllowedDeploymentAppTypes[k] = v
-		}
-		if !impl.deploymentConfig.IsInternalUse {
-			if isGitOpsConfigured && AllowedDeploymentAppTypes[util.PIPELINE_DEPLOYMENT_TYPE_ACD] {
-				pipeline.DeploymentAppType = util.PIPELINE_DEPLOYMENT_TYPE_ACD
-			} else if AllowedDeploymentAppTypes[util.PIPELINE_DEPLOYMENT_TYPE_HELM] {
-				pipeline.DeploymentAppType = util.PIPELINE_DEPLOYMENT_TYPE_HELM
-			}
-		}
-		if pipeline.DeploymentAppType == "" {
-			if isGitOpsConfigured && AllowedDeploymentAppTypes[util.PIPELINE_DEPLOYMENT_TYPE_ACD] {
-				pipeline.DeploymentAppType = util.PIPELINE_DEPLOYMENT_TYPE_ACD
-			} else if AllowedDeploymentAppTypes[util.PIPELINE_DEPLOYMENT_TYPE_HELM] {
-				pipeline.DeploymentAppType = util.PIPELINE_DEPLOYMENT_TYPE_HELM
-			}
-		}
-	}
-}
-
-func (impl *PipelineBuilderImpl) CreateCdPipelines(pipelineCreateRequest *bean.CdPipelines, ctx context.Context) (*bean.CdPipelines, error) {
-
-	//Validation for checking deployment App type
-	isGitOpsConfigured, err := impl.IsGitopsConfigured()
-
-	for _, pipeline := range pipelineCreateRequest.Pipelines {
-		// if no deployment app type sent from user then we'll not validate
-		deploymentConfig, err := impl.GetDeploymentConfigMap(pipeline.EnvironmentId)
-		if err != nil {
-			return nil, err
-		}
-		impl.SetPipelineDeploymentAppType(pipelineCreateRequest, isGitOpsConfigured, deploymentConfig)
-		if err := impl.validateDeploymentAppType(pipeline, deploymentConfig); err != nil {
-			impl.logger.Errorw("validation error in creating pipeline", "name", pipeline.Name, "err", err)
-			return nil, err
-		}
-	}
-
-	isGitOpsRequiredForCD := impl.IsGitOpsRequiredForCD(pipelineCreateRequest)
-	app, err := impl.appRepo.FindById(pipelineCreateRequest.AppId)
-	if err != nil {
-		impl.logger.Errorw("app not found", "err", err, "appId", pipelineCreateRequest.AppId)
-		return nil, err
-	}
-	_, err = impl.ValidateCDPipelineRequest(pipelineCreateRequest, isGitOpsConfigured, isGitOpsRequiredForCD)
-	if err != nil {
-		return nil, err
-	}
-
-	// TODO: creating git repo for all apps irrespective of acd or helm
-	if isGitOpsConfigured && isGitOpsRequiredForCD {
-
-		gitopsRepoName, chartGitAttr, err := impl.appService.CreateGitopsRepo(app, pipelineCreateRequest.UserId)
-		if err != nil {
-			impl.logger.Errorw("error in creating git repo", "err", err)
-			return nil, err
-		}
-
-		err = impl.RegisterInACD(gitopsRepoName, chartGitAttr, pipelineCreateRequest.UserId, ctx)
-		if err != nil {
-			impl.logger.Errorw("error in registering app in acd", "err", err)
-			return nil, err
-		}
-
-		err = impl.updateGitRepoUrlInCharts(pipelineCreateRequest.AppId, chartGitAttr, pipelineCreateRequest.UserId)
-		if err != nil {
-			impl.logger.Errorw("error in updating git repo url in charts", "err", err)
-			return nil, err
-		}
-	}
-
-	for _, pipeline := range pipelineCreateRequest.Pipelines {
-
-		id, err := impl.createCdPipeline(ctx, app, pipeline, pipelineCreateRequest.UserId)
-		if err != nil {
-			impl.logger.Errorw("error in creating pipeline", "name", pipeline.Name, "err", err)
-			return nil, err
-		}
-		pipeline.Id = id
-
-		//creating pipeline_stage entry here after tx commit due to FK issue
-		if pipeline.PreDeployStage != nil && len(pipeline.PreDeployStage.Steps) > 0 {
-			err = impl.pipelineStageService.CreatePipelineStage(pipeline.PreDeployStage, repository5.PIPELINE_STAGE_TYPE_PRE_CD, id, pipelineCreateRequest.UserId)
-			if err != nil {
-				impl.logger.Errorw("error in creating pre-cd stage", "err", err, "preCdStage", pipeline.PreDeployStage, "pipelineId", id)
-				return nil, err
-			}
-		}
-		if pipeline.PostDeployStage != nil && len(pipeline.PostDeployStage.Steps) > 0 {
-			err = impl.pipelineStageService.CreatePipelineStage(pipeline.PostDeployStage, repository5.PIPELINE_STAGE_TYPE_POST_CD, id, pipelineCreateRequest.UserId)
-			if err != nil {
-				impl.logger.Errorw("error in creating post-cd stage", "err", err, "postCdStage", pipeline.PostDeployStage, "pipelineId", id)
-				return nil, err
-			}
-		}
-
-	}
-
-	return pipelineCreateRequest, nil
-}
-
 func (impl *PipelineBuilderImpl) validateDeploymentAppType(pipeline *bean.CDPipelineConfigObject, deploymentConfig map[string]bool) error {
 
 	// Config value doesn't exist in attribute table
@@ -2170,286 +967,6 @@ func validDeploymentConfigReceived(deploymentConfig map[string]bool, deploymentT
 		}
 	}
 	return false
-}
-
-func (impl *PipelineBuilderImpl) GetDeploymentConfigMap(environmentId int) (map[string]bool, error) {
-	var deploymentConfig map[string]map[string]bool
-	var deploymentConfigEnv map[string]bool
-	deploymentConfigValues, err := impl.attributesRepository.FindByKey(attributes.ENFORCE_DEPLOYMENT_TYPE_CONFIG)
-	if err == pg.ErrNoRows {
-		return deploymentConfigEnv, nil
-	}
-	//if empty config received(doesn't exist in table) which can't be parsed
-	if deploymentConfigValues.Value != "" {
-		if err := json.Unmarshal([]byte(deploymentConfigValues.Value), &deploymentConfig); err != nil {
-			rerr := &util.ApiError{
-				HttpStatusCode:  http.StatusInternalServerError,
-				InternalMessage: err.Error(),
-				UserMessage:     "Failed to fetch deployment config values from the attributes table",
-			}
-			return deploymentConfigEnv, rerr
-		}
-		deploymentConfigEnv, _ = deploymentConfig[fmt.Sprintf("%d", environmentId)]
-	}
-	return deploymentConfigEnv, nil
-}
-
-func (impl *PipelineBuilderImpl) PatchCdPipelines(cdPipelines *bean.CDPatchRequest, ctx context.Context) (*bean.CdPipelines, error) {
-	pipelineRequest := &bean.CdPipelines{
-		UserId:    cdPipelines.UserId,
-		AppId:     cdPipelines.AppId,
-		Pipelines: []*bean.CDPipelineConfigObject{cdPipelines.Pipeline},
-	}
-	deleteAction := bean.CASCADE_DELETE
-	if cdPipelines.ForceDelete {
-		deleteAction = bean.FORCE_DELETE
-	} else if cdPipelines.NonCascadeDelete {
-		deleteAction = bean.NON_CASCADE_DELETE
-	}
-	switch cdPipelines.Action {
-	case bean.CD_CREATE:
-		return impl.CreateCdPipelines(pipelineRequest, ctx)
-	case bean.CD_UPDATE:
-		err := impl.updateCdPipeline(ctx, cdPipelines.Pipeline, cdPipelines.UserId)
-		return pipelineRequest, err
-	case bean.CD_DELETE:
-		pipeline, err := impl.pipelineRepository.FindById(cdPipelines.Pipeline.Id)
-		if err != nil {
-			impl.logger.Errorw("error in getting cd pipeline by id", "err", err, "id", cdPipelines.Pipeline.Id)
-			return pipelineRequest, err
-		}
-		deleteResponse, err := impl.DeleteCdPipeline(pipeline, ctx, deleteAction, false, cdPipelines.UserId)
-		pipelineRequest.AppDeleteResponse = deleteResponse
-		return pipelineRequest, err
-	case bean.CD_DELETE_PARTIAL:
-		pipeline, err := impl.pipelineRepository.FindById(cdPipelines.Pipeline.Id)
-		if err != nil {
-			impl.logger.Errorw("error in getting cd pipeline by id", "err", err, "id", cdPipelines.Pipeline.Id)
-			return pipelineRequest, err
-		}
-		deleteResponse, err := impl.DeleteCdPipelinePartial(pipeline, ctx, deleteAction, cdPipelines.UserId)
-		pipelineRequest.AppDeleteResponse = deleteResponse
-		return pipelineRequest, err
-	default:
-		return nil, &util.ApiError{Code: "404", HttpStatusCode: 404, UserMessage: "operation not supported"}
-	}
-}
-
-func (impl *PipelineBuilderImpl) DeleteCdPipeline(pipeline *pipelineConfig.Pipeline, ctx context.Context, deleteAction int, deleteFromAcd bool, userId int32) (*bean.AppDeleteResponseDTO, error) {
-	cascadeDelete := true
-	forceDelete := false
-	deleteResponse := &bean.AppDeleteResponseDTO{
-		DeleteInitiated:  false,
-		ClusterReachable: true,
-	}
-	if deleteAction == bean.FORCE_DELETE {
-		forceDelete = true
-		cascadeDelete = false
-	} else if deleteAction == bean.NON_CASCADE_DELETE {
-		cascadeDelete = false
-	}
-	// updating cluster reachable flag
-	clusterBean, err := impl.clusterRepository.FindById(pipeline.Environment.ClusterId)
-	if err != nil {
-		impl.logger.Errorw("error in getting cluster details", "err", err, "clusterId", pipeline.Environment.ClusterId)
-	}
-	deleteResponse.ClusterName = clusterBean.ClusterName
-	if len(clusterBean.ErrorInConnecting) > 0 {
-		deleteResponse.ClusterReachable = false
-	}
-	//getting children CD pipeline details
-	childNodes, err := impl.appWorkflowRepository.FindWFCDMappingByParentCDPipelineId(pipeline.Id)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("error in getting children cd details", "err", err)
-		return deleteResponse, err
-	} else if len(childNodes) > 0 {
-		impl.logger.Debugw("cannot delete cd pipeline, contains children cd")
-		return deleteResponse, fmt.Errorf("Please delete children CD pipelines before deleting this pipeline.")
-	}
-	//getting deployment group for this pipeline
-	deploymentGroupNames, err := impl.deploymentGroupRepository.GetNamesByAppIdAndEnvId(pipeline.EnvironmentId, pipeline.AppId)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("error in getting deployment group names by appId and envId", "err", err)
-		return deleteResponse, err
-	} else if len(deploymentGroupNames) > 0 {
-		groupNamesByte, err := json.Marshal(deploymentGroupNames)
-		if err != nil {
-			impl.logger.Errorw("error in marshaling deployment group names", "err", err, "deploymentGroupNames", deploymentGroupNames)
-		}
-		impl.logger.Debugw("cannot delete cd pipeline, is being used in deployment group")
-		return deleteResponse, fmt.Errorf("Please remove this CD pipeline from deployment groups : %s", string(groupNamesByte))
-	}
-	dbConnection := impl.pipelineRepository.GetConnection()
-	tx, err := dbConnection.Begin()
-	if err != nil {
-		return deleteResponse, err
-	}
-	// Rollback tx on error.
-	defer tx.Rollback()
-	if err = impl.ciCdPipelineOrchestrator.DeleteCdPipeline(pipeline.Id, userId, tx); err != nil {
-		impl.logger.Errorw("err in deleting pipeline from db", "id", pipeline, "err", err)
-		return deleteResponse, err
-	}
-	// delete entry in app_status table
-	err = impl.appStatusRepository.Delete(tx, pipeline.AppId, pipeline.EnvironmentId)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("err in deleting app_status from db", "appId", pipeline.AppId, "envId", pipeline.EnvironmentId, "err", err)
-		return deleteResponse, err
-	}
-	//delete app workflow mapping
-	appWorkflowMapping, err := impl.appWorkflowRepository.FindWFCDMappingByCDPipelineId(pipeline.Id)
-	if err != nil {
-		impl.logger.Errorw("error in deleting workflow mapping", "err", err)
-		return deleteResponse, err
-	}
-	if appWorkflowMapping.ParentType == appWorkflow.WEBHOOK {
-		childNodes, err := impl.appWorkflowRepository.FindWFCDMappingByExternalCiId(appWorkflowMapping.ParentId)
-		if err != nil && !util.IsErrNoRows(err) {
-			impl.logger.Errorw("error in fetching external ci", "err", err)
-			return deleteResponse, err
-		}
-		noOtherChildNodes := true
-		for _, childNode := range childNodes {
-			if appWorkflowMapping.Id != childNode.Id {
-				noOtherChildNodes = false
-			}
-		}
-		if noOtherChildNodes {
-			externalCiPipeline, err := impl.ciPipelineRepository.FindExternalCiById(appWorkflowMapping.ParentId)
-			if err != nil {
-				impl.logger.Errorw("error in deleting workflow mapping", "err", err)
-				return deleteResponse, err
-			}
-			externalCiPipeline.Active = false
-			externalCiPipeline.UpdatedOn = time.Now()
-			externalCiPipeline.UpdatedBy = userId
-			_, err = impl.ciPipelineRepository.UpdateExternalCi(externalCiPipeline, tx)
-			if err != nil {
-				impl.logger.Errorw("error in deleting workflow mapping", "err", err)
-				return deleteResponse, err
-			}
-
-			appWorkflow, err := impl.appWorkflowRepository.FindById(appWorkflowMapping.AppWorkflowId)
-			if err != nil {
-				impl.logger.Errorw("error in deleting workflow mapping", "err", err)
-				return deleteResponse, err
-			}
-			err = impl.appWorkflowRepository.DeleteAppWorkflow(appWorkflow, tx)
-			if err != nil {
-				impl.logger.Errorw("error in deleting workflow mapping", "err", err)
-				return deleteResponse, err
-			}
-		}
-	}
-	appWorkflowMapping.UpdatedBy = userId
-	appWorkflowMapping.UpdatedOn = time.Now()
-	err = impl.appWorkflowRepository.DeleteAppWorkflowMapping(appWorkflowMapping, tx)
-	if err != nil {
-		impl.logger.Errorw("error in deleting workflow mapping", "err", err)
-		return deleteResponse, err
-	}
-
-	if pipeline.PreStageConfig != "" {
-		err = impl.prePostCdScriptHistoryService.CreatePrePostCdScriptHistory(pipeline, tx, repository4.PRE_CD_TYPE, false, 0, time.Time{})
-		if err != nil {
-			impl.logger.Errorw("error in creating pre cd script entry", "err", err, "pipeline", pipeline)
-			return deleteResponse, err
-		}
-	}
-	if pipeline.PostStageConfig != "" {
-		err = impl.prePostCdScriptHistoryService.CreatePrePostCdScriptHistory(pipeline, tx, repository4.POST_CD_TYPE, false, 0, time.Time{})
-		if err != nil {
-			impl.logger.Errorw("error in creating post cd script entry", "err", err, "pipeline", pipeline)
-			return deleteResponse, err
-		}
-	}
-	cdPipelinePluginDeleteReq, err := impl.GetCdPipelineById(pipeline.Id)
-	if err != nil {
-		impl.logger.Errorw("error in getting cdPipeline by id", "err", err, "id", pipeline.Id)
-		return deleteResponse, err
-	}
-	if cdPipelinePluginDeleteReq.PreDeployStage != nil && cdPipelinePluginDeleteReq.PreDeployStage.Id > 0 {
-		//deleting pre-stage
-		err = impl.pipelineStageService.DeletePipelineStage(cdPipelinePluginDeleteReq.PreDeployStage, userId, tx)
-		if err != nil {
-			impl.logger.Errorw("error in deleting pre-CD stage", "err", err, "preDeployStage", cdPipelinePluginDeleteReq.PreDeployStage)
-			return deleteResponse, err
-		}
-	}
-	if cdPipelinePluginDeleteReq.PostDeployStage != nil && cdPipelinePluginDeleteReq.PostDeployStage.Id > 0 {
-		//deleting post-stage
-		err = impl.pipelineStageService.DeletePipelineStage(cdPipelinePluginDeleteReq.PostDeployStage, userId, tx)
-		if err != nil {
-			impl.logger.Errorw("error in deleting post-CD stage", "err", err, "postDeployStage", cdPipelinePluginDeleteReq.PostDeployStage)
-			return deleteResponse, err
-		}
-	}
-	//delete app from argo cd, if created
-	if pipeline.DeploymentAppCreated == true {
-		deploymentAppName := fmt.Sprintf("%s-%s", pipeline.App.AppName, pipeline.Environment.Name)
-		if util.IsAcdApp(pipeline.DeploymentAppType) {
-			if !deleteResponse.ClusterReachable {
-				impl.logger.Errorw("cluster connection error", "err", clusterBean.ErrorInConnecting)
-				if cascadeDelete {
-					return deleteResponse, nil
-				}
-			}
-			impl.logger.Debugw("acd app is already deleted for this pipeline", "pipeline", pipeline)
-			if deleteFromAcd {
-				req := &application2.ApplicationDeleteRequest{
-					Name:    &deploymentAppName,
-					Cascade: &cascadeDelete,
-				}
-				if _, err := impl.application.Delete(ctx, req); err != nil {
-					impl.logger.Errorw("err in deleting pipeline on argocd", "id", pipeline, "err", err)
-
-					if forceDelete {
-						impl.logger.Warnw("error while deletion of app in acd, continue to delete in db as this operation is force delete", "error", err)
-					} else {
-						//statusError, _ := err.(*errors2.StatusError)
-						if cascadeDelete && strings.Contains(err.Error(), "code = NotFound") {
-							err = &util.ApiError{
-								UserMessage:     "Could not delete as application not found in argocd",
-								InternalMessage: err.Error(),
-							}
-						} else {
-							err = &util.ApiError{
-								UserMessage:     "Could not delete application",
-								InternalMessage: err.Error(),
-							}
-						}
-						return deleteResponse, err
-					}
-				}
-				impl.logger.Infow("app deleted from argocd", "id", pipeline.Id, "pipelineName", pipeline.Name, "app", deploymentAppName)
-			}
-		} else if util.IsHelmApp(pipeline.DeploymentAppType) {
-			appIdentifier := &client.AppIdentifier{
-				ClusterId:   pipeline.Environment.ClusterId,
-				ReleaseName: deploymentAppName,
-				Namespace:   pipeline.Environment.Namespace,
-			}
-			deleteResourceResponse, err := impl.helmAppService.DeleteApplication(ctx, appIdentifier)
-			if forceDelete {
-				impl.logger.Warnw("error while deletion of helm application, ignore error and delete from db since force delete req", "error", err, "pipelineId", pipeline.Id)
-			} else {
-				if err != nil {
-					impl.logger.Errorw("error in deleting helm application", "error", err, "appIdentifier", appIdentifier)
-					return deleteResponse, err
-				}
-				if deleteResourceResponse == nil || !deleteResourceResponse.GetSuccess() {
-					return deleteResponse, errors.New("delete application response unsuccessful")
-				}
-			}
-		}
-	}
-	err = tx.Commit()
-	if err != nil {
-		impl.logger.Errorw("error in committing db transaction", "err", err)
-		return deleteResponse, err
-	}
-	deleteResponse.DeleteInitiated = true
-	return deleteResponse, nil
 }
 
 func (impl *PipelineBuilderImpl) DeleteCdPipelinePartial(pipeline *pipelineConfig.Pipeline, ctx context.Context, deleteAction int, userId int32) (*bean.AppDeleteResponseDTO, error) {
@@ -2559,476 +1076,6 @@ func (impl *PipelineBuilderImpl) DeleteCdPipelinePartial(pipeline *pipelineConfi
 	}
 	return deleteResponse, nil
 }
-
-func (impl *PipelineBuilderImpl) DeleteACDAppCdPipelineWithNonCascade(pipeline *pipelineConfig.Pipeline, ctx context.Context, forceDelete bool, userId int32) error {
-	if forceDelete {
-		_, err := impl.DeleteCdPipeline(pipeline, ctx, bean.FORCE_DELETE, false, userId)
-		return err
-	}
-	//delete app from argo cd with non-cascade, if created
-	if pipeline.DeploymentAppCreated && util.IsAcdApp(pipeline.DeploymentAppType) {
-		appDetails, err := impl.appRepo.FindById(pipeline.AppId)
-		deploymentAppName := fmt.Sprintf("%s-%s", appDetails.AppName, pipeline.Environment.Name)
-		impl.logger.Debugw("acd app is already deleted for this pipeline", "pipeline", pipeline)
-		cascadeDelete := false
-		req := &application2.ApplicationDeleteRequest{
-			Name:    &deploymentAppName,
-			Cascade: &cascadeDelete,
-		}
-		if _, err = impl.application.Delete(ctx, req); err != nil {
-			impl.logger.Errorw("err in deleting pipeline on argocd", "id", pipeline, "err", err)
-			//statusError, _ := err.(*errors2.StatusError)
-			if !strings.Contains(err.Error(), "code = NotFound") {
-				err = &util.ApiError{
-					UserMessage:     "Could not delete application",
-					InternalMessage: err.Error(),
-				}
-				return err
-			}
-		}
-
-	}
-	return nil
-}
-
-func (impl *PipelineBuilderImpl) ChangeDeploymentType(ctx context.Context,
-	request *bean.DeploymentAppTypeChangeRequest) (*bean.DeploymentAppTypeChangeResponse, error) {
-
-	var response *bean.DeploymentAppTypeChangeResponse
-	var deleteDeploymentType bean.DeploymentType
-	var err error
-
-	if request.DesiredDeploymentType == bean.ArgoCd {
-		deleteDeploymentType = bean.Helm
-	} else {
-		deleteDeploymentType = bean.ArgoCd
-	}
-
-	// Force delete apps
-	response, err = impl.DeleteDeploymentAppsForEnvironment(ctx,
-		request.EnvId, deleteDeploymentType, request.ExcludeApps, request.IncludeApps, request.UserId)
-
-	if err != nil {
-		return nil, err
-	}
-
-	// Updating the env id and desired deployment app type received from request in the response
-	response.EnvId = request.EnvId
-	response.DesiredDeploymentType = request.DesiredDeploymentType
-	response.TriggeredPipelines = make([]*bean.CdPipelineTrigger, 0)
-
-	// Update the deployment app type to Helm and toggle deployment_app_created to false in db
-	var cdPipelineIds []int
-	for _, item := range response.SuccessfulPipelines {
-		cdPipelineIds = append(cdPipelineIds, item.Id)
-	}
-
-	// If nothing to update in db
-	if len(cdPipelineIds) == 0 {
-		return response, nil
-	}
-
-	// Update in db
-	err = impl.pipelineRepository.UpdateCdPipelineDeploymentAppInFilter(string(request.DesiredDeploymentType),
-		cdPipelineIds, request.UserId, false, true)
-
-	if err != nil {
-		impl.logger.Errorw("failed to update deployment app type in db",
-			"pipeline ids", cdPipelineIds,
-			"desired deployment type", request.DesiredDeploymentType,
-			"err", err)
-
-		return response, nil
-	}
-
-	if !request.AutoTriggerDeployment {
-		return response, nil
-	}
-
-	// Bulk trigger all the successfully changed pipelines (async)
-	bulkTriggerRequest := make([]*BulkTriggerRequest, 0)
-
-	pipelineIds := make([]int, 0, len(response.SuccessfulPipelines))
-	for _, item := range response.SuccessfulPipelines {
-		pipelineIds = append(pipelineIds, item.Id)
-	}
-
-	// Get all pipelines
-	pipelines, err := impl.pipelineRepository.FindByIdsIn(pipelineIds)
-	if err != nil {
-		impl.logger.Errorw("failed to fetch pipeline details",
-			"ids", pipelineIds,
-			"err", err)
-
-		return response, nil
-	}
-
-	for _, pipeline := range pipelines {
-
-		artifactDetails, err := impl.RetrieveArtifactsByCDPipeline(pipeline, "DEPLOY")
-
-		if err != nil {
-			impl.logger.Errorw("failed to fetch artifact details for cd pipeline",
-				"pipelineId", pipeline.Id,
-				"appId", pipeline.AppId,
-				"envId", pipeline.EnvironmentId,
-				"err", err)
-
-			return response, nil
-		}
-
-		if artifactDetails.LatestWfArtifactId == 0 || artifactDetails.LatestWfArtifactStatus == "" {
-			continue
-		}
-
-		bulkTriggerRequest = append(bulkTriggerRequest, &BulkTriggerRequest{
-			CiArtifactId: artifactDetails.LatestWfArtifactId,
-			PipelineId:   pipeline.Id,
-		})
-		response.TriggeredPipelines = append(response.TriggeredPipelines, &bean.CdPipelineTrigger{
-			CiArtifactId: artifactDetails.LatestWfArtifactId,
-			PipelineId:   pipeline.Id,
-		})
-	}
-
-	// pg panics if empty slice is passed as an argument
-	if len(bulkTriggerRequest) == 0 {
-		return response, nil
-	}
-
-	// Trigger
-	_, err = impl.workflowDagExecutor.TriggerBulkDeploymentAsync(bulkTriggerRequest, request.UserId)
-
-	if err != nil {
-		impl.logger.Errorw("failed to bulk trigger cd pipelines with error: "+err.Error(),
-			"err", err)
-	}
-	return response, nil
-}
-
-func (impl *PipelineBuilderImpl) ChangePipelineDeploymentType(ctx context.Context,
-	request *bean.DeploymentAppTypeChangeRequest) (*bean.DeploymentAppTypeChangeResponse, error) {
-
-	response := &bean.DeploymentAppTypeChangeResponse{
-		EnvId:                 request.EnvId,
-		DesiredDeploymentType: request.DesiredDeploymentType,
-		TriggeredPipelines:    make([]*bean.CdPipelineTrigger, 0),
-	}
-
-	var deleteDeploymentType bean.DeploymentType
-
-	if request.DesiredDeploymentType == bean.ArgoCd {
-		deleteDeploymentType = bean.Helm
-	} else {
-		deleteDeploymentType = bean.ArgoCd
-	}
-
-	pipelines, err := impl.pipelineRepository.FindActiveByEnvIdAndDeploymentType(request.EnvId,
-		string(deleteDeploymentType), request.ExcludeApps, request.IncludeApps)
-
-	if err != nil {
-		impl.logger.Errorw("Error fetching cd pipelines",
-			"environmentId", request.EnvId,
-			"currentDeploymentAppType", string(deleteDeploymentType),
-			"err", err)
-		return response, err
-	}
-
-	var pipelineIds []int
-	for _, item := range pipelines {
-		pipelineIds = append(pipelineIds, item.Id)
-	}
-
-	if len(pipelineIds) == 0 {
-		return response, nil
-	}
-
-	err = impl.pipelineRepository.UpdateCdPipelineDeploymentAppInFilter(string(request.DesiredDeploymentType),
-		pipelineIds, request.UserId, false, true)
-
-	if err != nil {
-		impl.logger.Errorw("failed to update deployment app type in db",
-			"pipeline ids", pipelineIds,
-			"desired deployment type", request.DesiredDeploymentType,
-			"err", err)
-
-		return response, nil
-	}
-	deleteResponse := impl.DeleteDeploymentApps(ctx, pipelines, request.UserId)
-
-	response.SuccessfulPipelines = deleteResponse.SuccessfulPipelines
-	response.FailedPipelines = deleteResponse.FailedPipelines
-
-	var cdPipelineIds []int
-	for _, item := range response.FailedPipelines {
-		cdPipelineIds = append(cdPipelineIds, item.Id)
-	}
-
-	if len(cdPipelineIds) == 0 {
-		return response, nil
-	}
-
-	err = impl.pipelineRepository.UpdateCdPipelineDeploymentAppInFilter(string(deleteDeploymentType),
-		cdPipelineIds, request.UserId, true, false)
-
-	if err != nil {
-		impl.logger.Errorw("failed to update deployment app type in db",
-			"pipeline ids", cdPipelineIds,
-			"desired deployment type", request.DesiredDeploymentType,
-			"err", err)
-
-		return response, nil
-	}
-
-	return response, nil
-}
-
-func (impl *PipelineBuilderImpl) TriggerDeploymentAfterTypeChange(ctx context.Context,
-	request *bean.DeploymentAppTypeChangeRequest) (*bean.DeploymentAppTypeChangeResponse, error) {
-
-	response := &bean.DeploymentAppTypeChangeResponse{
-		EnvId:                 request.EnvId,
-		DesiredDeploymentType: request.DesiredDeploymentType,
-		TriggeredPipelines:    make([]*bean.CdPipelineTrigger, 0),
-	}
-	var err error
-
-	cdPipelines, err := impl.pipelineRepository.FindActiveByEnvIdAndDeploymentType(request.EnvId,
-		string(request.DesiredDeploymentType), request.ExcludeApps, request.IncludeApps)
-
-	if err != nil {
-		impl.logger.Errorw("Error fetching cd pipelines",
-			"environmentId", request.EnvId,
-			"desiredDeploymentAppType", string(request.DesiredDeploymentType),
-			"err", err)
-		return response, err
-	}
-
-	var cdPipelineIds []int
-	for _, item := range cdPipelines {
-		cdPipelineIds = append(cdPipelineIds, item.Id)
-	}
-
-	if len(cdPipelineIds) == 0 {
-		return response, nil
-	}
-
-	deleteResponse := impl.FetchDeletedApp(ctx, cdPipelines)
-
-	response.SuccessfulPipelines = deleteResponse.SuccessfulPipelines
-	response.FailedPipelines = deleteResponse.FailedPipelines
-
-	var successPipelines []int
-	for _, item := range response.SuccessfulPipelines {
-		successPipelines = append(successPipelines, item.Id)
-	}
-
-	bulkTriggerRequest := make([]*BulkTriggerRequest, 0)
-
-	pipelineIds := make([]int, 0, len(response.SuccessfulPipelines))
-	for _, item := range response.SuccessfulPipelines {
-		pipelineIds = append(pipelineIds, item.Id)
-	}
-
-	pipelines, err := impl.pipelineRepository.FindByIdsIn(pipelineIds)
-	if err != nil {
-		impl.logger.Errorw("failed to fetch pipeline details",
-			"ids", pipelineIds,
-			"err", err)
-
-		return response, nil
-	}
-
-	for _, pipeline := range pipelines {
-
-		artifactDetails, err := impl.RetrieveArtifactsByCDPipeline(pipeline, "DEPLOY")
-
-		if err != nil {
-			impl.logger.Errorw("failed to fetch artifact details for cd pipeline",
-				"pipelineId", pipeline.Id,
-				"appId", pipeline.AppId,
-				"envId", pipeline.EnvironmentId,
-				"err", err)
-
-			return response, nil
-		}
-
-		if artifactDetails.LatestWfArtifactId == 0 || artifactDetails.LatestWfArtifactStatus == "" {
-			continue
-		}
-
-		bulkTriggerRequest = append(bulkTriggerRequest, &BulkTriggerRequest{
-			CiArtifactId: artifactDetails.LatestWfArtifactId,
-			PipelineId:   pipeline.Id,
-		})
-		response.TriggeredPipelines = append(response.TriggeredPipelines, &bean.CdPipelineTrigger{
-			CiArtifactId: artifactDetails.LatestWfArtifactId,
-			PipelineId:   pipeline.Id,
-		})
-	}
-
-	if len(bulkTriggerRequest) == 0 {
-		return response, nil
-	}
-
-	_, err = impl.workflowDagExecutor.TriggerBulkDeploymentAsync(bulkTriggerRequest, request.UserId)
-
-	if err != nil {
-		impl.logger.Errorw("failed to bulk trigger cd pipelines with error: "+err.Error(),
-			"err", err)
-	}
-
-	err = impl.pipelineRepository.UpdateCdPipelineAfterDeployment(string(request.DesiredDeploymentType),
-		successPipelines, request.UserId, false)
-
-	if err != nil {
-		impl.logger.Errorw("failed to update cd pipelines with error: : "+err.Error(),
-			"err", err)
-	}
-
-	return response, nil
-}
-
-// DeleteDeploymentAppsForEnvironment takes in environment id and current deployment app type
-// and deletes all the cd pipelines for that deployment type in all apps that belongs to
-// that environment.
-func (impl *PipelineBuilderImpl) DeleteDeploymentAppsForEnvironment(ctx context.Context, environmentId int,
-	currentDeploymentAppType bean.DeploymentType, exclusionList []int, includeApps []int, userId int32) (*bean.DeploymentAppTypeChangeResponse, error) {
-
-	// fetch active pipelines from database for the given environment id and current deployment app type
-	pipelines, err := impl.pipelineRepository.FindActiveByEnvIdAndDeploymentType(environmentId,
-		string(currentDeploymentAppType), exclusionList, includeApps)
-
-	if err != nil {
-		impl.logger.Errorw("Error fetching cd pipelines",
-			"environmentId", environmentId,
-			"currentDeploymentAppType", currentDeploymentAppType,
-			"err", err)
-
-		return &bean.DeploymentAppTypeChangeResponse{
-			EnvId:               environmentId,
-			SuccessfulPipelines: []*bean.DeploymentChangeStatus{},
-			FailedPipelines:     []*bean.DeploymentChangeStatus{},
-		}, err
-	}
-
-	// Currently deleting apps only in argocd is supported
-	return impl.DeleteDeploymentApps(ctx, pipelines, userId), nil
-}
-
-// DeleteDeploymentApps takes in a list of pipelines and delete the applications
-
-func (impl *PipelineBuilderImpl) DeleteDeploymentApps(ctx context.Context,
-	pipelines []*pipelineConfig.Pipeline, userId int32) *bean.DeploymentAppTypeChangeResponse {
-
-	successfulPipelines := make([]*bean.DeploymentChangeStatus, 0)
-	failedPipelines := make([]*bean.DeploymentChangeStatus, 0)
-
-	isGitOpsConfigured, gitOpsConfigErr := impl.IsGitopsConfigured()
-
-	// Iterate over all the pipelines in the environment for given deployment app type
-	for _, pipeline := range pipelines {
-
-		var isValid bool
-		// check if pipeline info like app name and environment is empty or not
-		if failedPipelines, isValid = impl.isPipelineInfoValid(pipeline, failedPipelines); !isValid {
-			continue
-		}
-
-		var healthChkErr error
-		// check health of the app if it is argocd deployment type
-		if _, healthChkErr = impl.handleNotDeployedAppsIfArgoDeploymentType(pipeline, failedPipelines); healthChkErr != nil {
-
-			// cannot delete unhealthy app
-			continue
-		}
-
-		deploymentAppName := fmt.Sprintf("%s-%s", pipeline.App.AppName, pipeline.Environment.Name)
-		var err error
-
-		// delete request
-		if pipeline.DeploymentAppType == bean.ArgoCd {
-			err = impl.deleteArgoCdApp(ctx, pipeline, deploymentAppName, true)
-
-		} else {
-
-			// For converting from Helm to ArgoCD, GitOps should be configured
-			if gitOpsConfigErr != nil || !isGitOpsConfigured {
-				err = errors.New("GitOps not configured or unable to fetch GitOps configuration")
-
-			} else {
-				// Register app in ACD
-				var AcdRegisterErr, RepoURLUpdateErr error
-				gitopsRepoName, chartGitAttr, createGitRepoErr := impl.appService.CreateGitopsRepo(&app2.App{Id: pipeline.AppId, AppName: pipeline.App.AppName}, userId)
-				if createGitRepoErr != nil {
-					impl.logger.Errorw("error increating git repo", "err", err)
-				}
-				if createGitRepoErr == nil {
-					AcdRegisterErr = impl.RegisterInACD(gitopsRepoName,
-						chartGitAttr,
-						userId,
-						ctx)
-					if AcdRegisterErr != nil {
-						impl.logger.Errorw("error in registering acd app", "err", err)
-					}
-					if AcdRegisterErr == nil {
-						RepoURLUpdateErr = impl.chartTemplateService.UpdateGitRepoUrlInCharts(pipeline.AppId, chartGitAttr, userId)
-						if RepoURLUpdateErr != nil {
-							impl.logger.Errorw("error in updating git repo url in charts", "err", err)
-						}
-					}
-				}
-				if createGitRepoErr != nil {
-					err = createGitRepoErr
-				} else if AcdRegisterErr != nil {
-					err = AcdRegisterErr
-				} else if RepoURLUpdateErr != nil {
-					err = RepoURLUpdateErr
-				}
-			}
-			if err != nil {
-				impl.logger.Errorw("error registering app on ACD with error: "+err.Error(),
-					"deploymentAppName", deploymentAppName,
-					"envId", pipeline.EnvironmentId,
-					"appId", pipeline.AppId,
-					"err", err)
-
-				// deletion failed, append to the list of failed pipelines
-				failedPipelines = impl.handleFailedDeploymentAppChange(pipeline, failedPipelines,
-					"failed to register app on ACD with error: "+err.Error())
-
-				continue
-			}
-			err = impl.deleteHelmApp(ctx, pipeline)
-		}
-
-		if err != nil {
-			impl.logger.Errorw("error deleting app on "+pipeline.DeploymentAppType,
-				"deployment app name", deploymentAppName,
-				"err", err)
-
-			// deletion failed, append to the list of failed pipelines
-			failedPipelines = impl.handleFailedDeploymentAppChange(pipeline, failedPipelines,
-				"error deleting app with error: "+err.Error())
-
-			continue
-		}
-
-		// deletion successful, append to the list of successful pipelines
-		successfulPipelines = impl.appendToDeploymentChangeStatusList(
-			successfulPipelines,
-			pipeline,
-			"",
-			bean.INITIATED)
-	}
-
-	return &bean.DeploymentAppTypeChangeResponse{
-		SuccessfulPipelines: successfulPipelines,
-		FailedPipelines:     failedPipelines,
-	}
-}
-
-// from argocd / helm.
 
 func (impl *PipelineBuilderImpl) isGitRepoUrlPresent(appId int) bool {
 	fetchedChart, err := impl.chartRepository.FindLatestByAppId(appId)
@@ -3573,312 +1620,9 @@ func (impl *PipelineBuilderImpl) getStrategiesMapping(dbPipelineIds []int) (map[
 	return strategiesMapping, nil
 }
 
-func (impl *PipelineBuilderImpl) GetTriggerViewCdPipelinesForApp(appId int) (cdPipelines *bean.CdPipelines, err error) {
-	triggerViewCdPipelinesResp, err := impl.ciCdPipelineOrchestrator.GetCdPipelinesForApp(appId)
-	if err != nil {
-		impl.logger.Errorw("error in fetching triggerViewCdPipelinesResp by appId", "err", err, "appId", appId)
-		return triggerViewCdPipelinesResp, err
-	}
-	var dbPipelineIds []int
-	for _, dbPipeline := range triggerViewCdPipelinesResp.Pipelines {
-		dbPipelineIds = append(dbPipelineIds, dbPipeline.Id)
-	}
-
-	//construct strategiesMapping to get all strategies against pipelineId
-	strategiesMapping, err := impl.getStrategiesMapping(dbPipelineIds)
-	if err != nil {
-		return triggerViewCdPipelinesResp, err
-	}
-	for _, dbPipeline := range triggerViewCdPipelinesResp.Pipelines {
-		var strategies []*chartConfig.PipelineStrategy
-		var deploymentTemplate chartRepoRepository.DeploymentStrategy
-		if len(strategiesMapping[dbPipeline.Id]) != 0 {
-			strategies = strategiesMapping[dbPipeline.Id]
-		}
-		for _, item := range strategies {
-			if item.Default {
-				deploymentTemplate = item.Strategy
-			}
-		}
-		dbPipeline.DeploymentTemplate = deploymentTemplate
-	}
-
-	return triggerViewCdPipelinesResp, err
-}
-
-func (impl *PipelineBuilderImpl) GetCdPipelinesForApp(appId int) (cdPipelines *bean.CdPipelines, err error) {
-	cdPipelines, err = impl.ciCdPipelineOrchestrator.GetCdPipelinesForApp(appId)
-	if err != nil {
-		impl.logger.Errorw("error in fetching cd Pipelines for appId", "err", err, "appId", appId)
-		return nil, err
-	}
-	var envIds []*int
-	var dbPipelineIds []int
-	for _, dbPipeline := range cdPipelines.Pipelines {
-		envIds = append(envIds, &dbPipeline.EnvironmentId)
-		dbPipelineIds = append(dbPipelineIds, dbPipeline.Id)
-	}
-	if len(envIds) == 0 || len(dbPipelineIds) == 0 {
-		return cdPipelines, nil
-	}
-	envMapping := make(map[int]*repository2.Environment)
-	appWorkflowMapping := make(map[int]*appWorkflow.AppWorkflowMapping)
-
-	envs, err := impl.environmentRepository.FindByIds(envIds)
-	if err != nil && errors.IsNotFound(err) {
-		impl.logger.Errorw("error in fetching environments", "err", err)
-		return cdPipelines, err
-	}
-	//creating map for envId and respective env
-	for _, env := range envs {
-		envMapping[env.Id] = env
-	}
-	strategiesMapping, err := impl.getStrategiesMapping(dbPipelineIds)
-	if err != nil {
-		return cdPipelines, err
-	}
-	appWorkflowMappings, err := impl.appWorkflowRepository.FindByCDPipelineIds(dbPipelineIds)
-	if err != nil {
-		impl.logger.Errorw("error in fetching app workflow mappings by pipelineIds", "err", err)
-		return nil, err
-	}
-	for _, appWorkflow := range appWorkflowMappings {
-		appWorkflowMapping[appWorkflow.ComponentId] = appWorkflow
-	}
-
-	var pipelines []*bean.CDPipelineConfigObject
-	for _, dbPipeline := range cdPipelines.Pipelines {
-		environment := &repository2.Environment{}
-		var strategies []*chartConfig.PipelineStrategy
-		appToWorkflowMapping := &appWorkflow.AppWorkflowMapping{}
-
-		if envMapping[dbPipeline.EnvironmentId] != nil {
-			environment = envMapping[dbPipeline.EnvironmentId]
-		}
-		if len(strategiesMapping[dbPipeline.Id]) != 0 {
-			strategies = strategiesMapping[dbPipeline.Id]
-		}
-		if appWorkflowMapping[dbPipeline.Id] != nil {
-			appToWorkflowMapping = appWorkflowMapping[dbPipeline.Id]
-		}
-		var strategiesBean []bean.Strategy
-		var deploymentTemplate chartRepoRepository.DeploymentStrategy
-		for _, item := range strategies {
-			strategiesBean = append(strategiesBean, bean.Strategy{
-				Config:             []byte(item.Config),
-				DeploymentTemplate: item.Strategy,
-				Default:            item.Default,
-			})
-
-			if item.Default {
-				deploymentTemplate = item.Strategy
-			}
-		}
-		pipeline := &bean.CDPipelineConfigObject{
-			Id:                            dbPipeline.Id,
-			Name:                          dbPipeline.Name,
-			EnvironmentId:                 dbPipeline.EnvironmentId,
-			EnvironmentName:               environment.Name,
-			Description:                   environment.Description,
-			CiPipelineId:                  dbPipeline.CiPipelineId,
-			DeploymentTemplate:            deploymentTemplate,
-			TriggerType:                   dbPipeline.TriggerType,
-			Strategies:                    strategiesBean,
-			PreStage:                      dbPipeline.PreStage,
-			PostStage:                     dbPipeline.PostStage,
-			PreStageConfigMapSecretNames:  dbPipeline.PreStageConfigMapSecretNames,
-			PostStageConfigMapSecretNames: dbPipeline.PostStageConfigMapSecretNames,
-			RunPreStageInEnv:              dbPipeline.RunPreStageInEnv,
-			RunPostStageInEnv:             dbPipeline.RunPostStageInEnv,
-			DeploymentAppType:             dbPipeline.DeploymentAppType,
-			ParentPipelineType:            appToWorkflowMapping.ParentType,
-			ParentPipelineId:              appToWorkflowMapping.ParentId,
-			DeploymentAppDeleteRequest:    dbPipeline.DeploymentAppDeleteRequest,
-			IsVirtualEnvironment:          dbPipeline.IsVirtualEnvironment,
-			PreDeployStage:                dbPipeline.PreDeployStage,
-			PostDeployStage:               dbPipeline.PostDeployStage,
-		}
-		pipelines = append(pipelines, pipeline)
-	}
-	cdPipelines.Pipelines = pipelines
-	return cdPipelines, err
-}
-
-func (impl *PipelineBuilderImpl) GetCdPipelinesForAppAndEnv(appId int, envId int) (cdPipelines *bean.CdPipelines, err error) {
-	return impl.ciCdPipelineOrchestrator.GetCdPipelinesForAppAndEnv(appId, envId)
-}
-
 type ConfigMapSecretsResponse struct {
 	Maps    []bean2.ConfigSecretMap `json:"maps"`
 	Secrets []bean2.ConfigSecretMap `json:"secrets"`
-}
-
-func (impl *PipelineBuilderImpl) FetchConfigmapSecretsForCdStages(appId, envId, cdPipelineId int) (ConfigMapSecretsResponse, error) {
-	configMapSecrets, err := impl.appService.GetConfigMapAndSecretJson(appId, envId, cdPipelineId)
-	if err != nil {
-		impl.logger.Errorw("error while fetching config secrets ", "err", err)
-		return ConfigMapSecretsResponse{}, err
-	}
-	existingConfigMapSecrets := ConfigMapSecretsResponse{}
-	err = json.Unmarshal([]byte(configMapSecrets), &existingConfigMapSecrets)
-	if err != nil {
-		impl.logger.Error(err)
-		return ConfigMapSecretsResponse{}, err
-	}
-	return existingConfigMapSecrets, nil
-}
-
-func (impl *PipelineBuilderImpl) RetrieveParentDetails(pipelineId int) (parentId int, parentType bean2.WorkflowType, err error) {
-
-	workflow, err := impl.appWorkflowRepository.GetParentDetailsByPipelineId(pipelineId)
-	if err != nil {
-		impl.logger.Errorw("failed to get parent component details",
-			"componentId", pipelineId,
-			"err", err)
-		return 0, "", err
-	}
-
-	if workflow.ParentType == appWorkflow.CDPIPELINE {
-		// workflow is of type CD, check for stage
-		parentPipeline, err := impl.pipelineRepository.GetPostStageConfigById(workflow.ParentId)
-		if err != nil {
-			impl.logger.Errorw("failed to get the post_stage_config_yaml",
-				"cdPipelineId", workflow.ParentId,
-				"err", err)
-			return 0, "", err
-		}
-
-		if len(parentPipeline.PostStageConfig) == 0 {
-			return workflow.ParentId, bean2.CD_WORKFLOW_TYPE_DEPLOY, nil
-		}
-		return workflow.ParentId, bean2.CD_WORKFLOW_TYPE_POST, nil
-
-	} else if workflow.ParentType == appWorkflow.WEBHOOK {
-		// For webhook type
-		return workflow.ParentId, bean2.WEBHOOK_WORKFLOW_TYPE, nil
-	}
-
-	return workflow.ParentId, bean2.CI_WORKFLOW_TYPE, nil
-}
-
-// RetrieveArtifactsByCDPipeline returns all the artifacts for the cd pipeline (pre / deploy / post)
-func (impl *PipelineBuilderImpl) RetrieveArtifactsByCDPipeline(pipeline *pipelineConfig.Pipeline, stage bean2.WorkflowType) (*bean.CiArtifactResponse, error) {
-
-	// retrieve parent details
-	parentId, parentType, err := impl.RetrieveParentDetails(pipeline.Id)
-	if err != nil {
-		impl.logger.Errorw("failed to retrieve parent details",
-			"cdPipelineId", pipeline.Id,
-			"err", err)
-		return nil, err
-	}
-
-	parentCdId := 0
-	if parentType == bean2.CD_WORKFLOW_TYPE_POST || (parentType == bean2.CD_WORKFLOW_TYPE_DEPLOY && stage != bean2.CD_WORKFLOW_TYPE_POST) {
-		// parentCdId is being set to store the artifact currently deployed on parent cd (if applicable).
-		// Parent component is CD only if parent type is POST/DEPLOY
-		parentCdId = parentId
-	}
-
-	if stage == bean2.CD_WORKFLOW_TYPE_DEPLOY && len(pipeline.PreStageConfig) > 0 {
-		// Parent type will be PRE for DEPLOY stage
-		parentId = pipeline.Id
-		parentType = bean2.CD_WORKFLOW_TYPE_PRE
-	}
-	if stage == bean2.CD_WORKFLOW_TYPE_POST {
-		// Parent type will be DEPLOY for POST stage
-		parentId = pipeline.Id
-		parentType = bean2.CD_WORKFLOW_TYPE_DEPLOY
-	}
-
-	// Build artifacts for cd stages
-	var ciArtifacts []bean.CiArtifactBean
-	ciArtifactsResponse := &bean.CiArtifactResponse{}
-
-	artifactMap := make(map[int]int)
-	limit := 10
-
-	ciArtifacts, artifactMap, latestWfArtifactId, latestWfArtifactStatus, err := impl.
-		BuildArtifactsForCdStage(pipeline.Id, stage, ciArtifacts, artifactMap, false, limit, parentCdId)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("error in getting artifacts for child cd stage", "err", err, "stage", stage)
-		return nil, err
-	}
-
-	ciArtifacts, err = impl.BuildArtifactsForParentStage(pipeline.Id, parentId, parentType, ciArtifacts, artifactMap, limit, parentCdId)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("error in getting artifacts for cd", "err", err, "parentStage", parentType, "stage", stage)
-		return nil, err
-	}
-
-	//sorting ci artifacts on the basis of creation time
-	if ciArtifacts != nil {
-		sort.SliceStable(ciArtifacts, func(i, j int) bool {
-			return ciArtifacts[i].Id > ciArtifacts[j].Id
-		})
-	}
-
-	artifactIds := make([]int, 0, len(ciArtifacts))
-	for _, artifact := range ciArtifacts {
-		artifactIds = append(artifactIds, artifact.Id)
-	}
-
-	artifacts, err := impl.ciArtifactRepository.GetArtifactParentCiAndWorkflowDetailsByIds(artifactIds)
-	if err != nil {
-		return ciArtifactsResponse, err
-	}
-	imageTagsDataMap, err := impl.imageTaggingService.GetTagsDataMapByAppId(pipeline.AppId)
-	if err != nil {
-		impl.logger.Errorw("error in getting image tagging data with appId", "err", err, "appId", pipeline.AppId)
-		return ciArtifactsResponse, err
-	}
-
-	imageCommentsDataMap, err := impl.imageTaggingService.GetImageCommentsDataMapByArtifactIds(artifactIds)
-	if err != nil {
-		impl.logger.Errorw("error in getting GetImageCommentsDataMapByArtifactIds", "err", err, "appId", pipeline.AppId, "artifactIds", artifactIds)
-		return ciArtifactsResponse, err
-	}
-
-	for i, artifact := range artifacts {
-		if imageTaggingResp := imageTagsDataMap[ciArtifacts[i].Id]; imageTaggingResp != nil {
-			ciArtifacts[i].ImageReleaseTags = imageTaggingResp
-		}
-		if imageCommentResp := imageCommentsDataMap[ciArtifacts[i].Id]; imageCommentResp != nil {
-			ciArtifacts[i].ImageComment = imageCommentResp
-		}
-
-		if artifact.ExternalCiPipelineId != 0 {
-			// if external webhook continue
-			continue
-		}
-
-		var ciWorkflow *pipelineConfig.CiWorkflow
-		if artifact.ParentCiArtifact != 0 {
-			ciWorkflow, err = impl.ciWorkflowRepository.FindLastTriggeredWorkflowGitTriggersByArtifactId(artifact.ParentCiArtifact)
-			if err != nil {
-				impl.logger.Errorw("error in getting ci_workflow for artifacts", "err", err, "artifact", artifact, "parentStage", parentType, "stage", stage)
-				return ciArtifactsResponse, err
-			}
-
-		} else {
-			ciWorkflow, err = impl.ciWorkflowRepository.FindCiWorkflowGitTriggersById(*artifact.WorkflowId)
-			if err != nil {
-				impl.logger.Errorw("error in getting ci_workflow for artifacts", "err", err, "artifact", artifact, "parentStage", parentType, "stage", stage)
-				return ciArtifactsResponse, err
-			}
-		}
-		ciArtifacts[i].CiConfigureSourceType = ciWorkflow.GitTriggers[ciWorkflow.CiPipelineId].CiConfigureSourceType
-		ciArtifacts[i].CiConfigureSourceValue = ciWorkflow.GitTriggers[ciWorkflow.CiPipelineId].CiConfigureSourceValue
-	}
-
-	ciArtifactsResponse.CdPipelineId = pipeline.Id
-	ciArtifactsResponse.LatestWfArtifactId = latestWfArtifactId
-	ciArtifactsResponse.LatestWfArtifactStatus = latestWfArtifactStatus
-	if ciArtifacts == nil {
-		ciArtifacts = []bean.CiArtifactBean{}
-	}
-	ciArtifactsResponse.CiArtifacts = ciArtifacts
-	return ciArtifactsResponse, nil
 }
 
 func (impl *PipelineBuilderImpl) BuildArtifactsForParentStage(cdPipelineId int, parentId int, parentType bean2.WorkflowType, ciArtifacts []bean.CiArtifactBean, artifactMap map[int]int, limit int, parentCdId int) ([]bean.CiArtifactBean, error) {
@@ -3992,83 +1736,6 @@ func (impl *PipelineBuilderImpl) BuildArtifactsForCIParent(cdPipelineId int, par
 	return ciArtifacts, nil
 }
 
-func (impl *PipelineBuilderImpl) FetchArtifactForRollback(cdPipelineId, appId, offset, limit int) (bean.CiArtifactResponse, error) {
-	var deployedCiArtifacts []bean.CiArtifactBean
-	var deployedCiArtifactsResponse bean.CiArtifactResponse
-
-	cdWfrs, err := impl.cdWorkflowRepository.FetchArtifactsByCdPipelineId(cdPipelineId, bean2.CD_WORKFLOW_TYPE_DEPLOY, offset, limit)
-	if err != nil {
-		impl.logger.Errorw("error in getting artifacts for rollback by cdPipelineId", "err", err, "cdPipelineId", cdPipelineId)
-		return deployedCiArtifactsResponse, err
-	}
-	var ids []int32
-	for _, item := range cdWfrs {
-		ids = append(ids, item.TriggeredBy)
-	}
-	userEmails := make(map[int32]string)
-	users, err := impl.userService.GetByIds(ids)
-	if err != nil {
-		impl.logger.Errorw("unable to fetch users by ids", "err", err, "ids", ids)
-	}
-	for _, item := range users {
-		userEmails[item.Id] = item.EmailId
-	}
-
-	imageTagsDataMap, err := impl.imageTaggingService.GetTagsDataMapByAppId(appId)
-	if err != nil {
-		impl.logger.Errorw("error in getting image tagging data with appId", "err", err, "appId", appId)
-		return deployedCiArtifactsResponse, err
-	}
-	artifactIds := make([]int, 0)
-
-	for _, cdWfr := range cdWfrs {
-		ciArtifact := &repository.CiArtifact{}
-		if cdWfr.CdWorkflow != nil && cdWfr.CdWorkflow.CiArtifact != nil {
-			ciArtifact = cdWfr.CdWorkflow.CiArtifact
-		}
-		if ciArtifact == nil {
-			continue
-		}
-		mInfo, err := parseMaterialInfo([]byte(ciArtifact.MaterialInfo), ciArtifact.DataSource)
-		if err != nil {
-			mInfo = []byte("[]")
-			impl.logger.Errorw("error in parsing ciArtifact material info", "err", err, "ciArtifact", ciArtifact)
-		}
-		userEmail := userEmails[cdWfr.TriggeredBy]
-		deployedCiArtifacts = append(deployedCiArtifacts, bean.CiArtifactBean{
-			Id:           ciArtifact.Id,
-			Image:        ciArtifact.Image,
-			MaterialInfo: mInfo,
-			DeployedTime: formatDate(cdWfr.StartedOn, bean.LayoutRFC3339),
-			WfrId:        cdWfr.Id,
-			DeployedBy:   userEmail,
-		})
-		artifactIds = append(artifactIds, ciArtifact.Id)
-	}
-	imageCommentsDataMap, err := impl.imageTaggingService.GetImageCommentsDataMapByArtifactIds(artifactIds)
-	if err != nil {
-		impl.logger.Errorw("error in getting GetImageCommentsDataMapByArtifactIds", "err", err, "appId", appId, "artifactIds", artifactIds)
-		return deployedCiArtifactsResponse, err
-	}
-
-	for i, _ := range deployedCiArtifacts {
-		if imageTaggingResp := imageTagsDataMap[deployedCiArtifacts[i].Id]; imageTaggingResp != nil {
-			deployedCiArtifacts[i].ImageReleaseTags = imageTaggingResp
-		}
-		if imageCommentResp := imageCommentsDataMap[deployedCiArtifacts[i].Id]; imageCommentResp != nil {
-			deployedCiArtifacts[i].ImageComment = imageCommentResp
-		}
-	}
-
-	deployedCiArtifactsResponse.CdPipelineId = cdPipelineId
-	if deployedCiArtifacts == nil {
-		deployedCiArtifacts = []bean.CiArtifactBean{}
-	}
-	deployedCiArtifactsResponse.CiArtifacts = deployedCiArtifacts
-
-	return deployedCiArtifactsResponse, nil
-}
-
 func parseMaterialInfo(materialInfo json.RawMessage, source string) (json.RawMessage, error) {
 	if source != "GOCD" && source != "CI-RUNNER" && source != "EXTERNAL" {
 		return nil, fmt.Errorf("datasource: %s not supported", source)
@@ -4118,40 +1785,6 @@ func parseMaterialInfo(materialInfo json.RawMessage, source string) (json.RawMes
 	return mInfo, err
 }
 
-func (impl *PipelineBuilderImpl) FindAppsByTeamId(teamId int) ([]*AppBean, error) {
-	var appsRes []*AppBean
-	apps, err := impl.appRepo.FindAppsByTeamId(teamId)
-	if err != nil {
-		impl.logger.Errorw("error while fetching app", "err", err)
-		return nil, err
-	}
-	for _, app := range apps {
-		appsRes = append(appsRes, &AppBean{Id: app.Id, Name: app.AppName})
-	}
-	return appsRes, err
-}
-
-func (impl *PipelineBuilderImpl) FindAppsByTeamName(teamName string) ([]AppBean, error) {
-	var appsRes []AppBean
-	apps, err := impl.appRepo.FindAppsByTeamName(teamName)
-	if err != nil {
-		impl.logger.Errorw("error while fetching app", "err", err)
-		return nil, err
-	}
-	for _, app := range apps {
-		appsRes = append(appsRes, AppBean{Id: app.Id, Name: app.AppName})
-	}
-	return appsRes, err
-}
-
-func (impl *PipelineBuilderImpl) FindPipelineById(cdPipelineId int) (*pipelineConfig.Pipeline, error) {
-	return impl.pipelineRepository.FindById(cdPipelineId)
-}
-
-func (impl *PipelineBuilderImpl) FindAppAndEnvDetailsByPipelineId(cdPipelineId int) (*pipelineConfig.Pipeline, error) {
-	return impl.pipelineRepository.FindAppAndEnvDetailsByPipelineId(cdPipelineId)
-}
-
 type TeamAppBean struct {
 	ProjectId   int        `json:"projectId"`
 	ProjectName string     `json:"projectName"`
@@ -4164,72 +1797,6 @@ type AppBean struct {
 	TeamId int    `json:"teamId,omitempty"`
 }
 
-func (impl *PipelineBuilderImpl) GetAppList() ([]AppBean, error) {
-	var appsRes []AppBean
-	apps, err := impl.appRepo.FindAll()
-	if err != nil {
-		impl.logger.Errorw("error while fetching app", "err", err)
-		return nil, err
-	}
-	for _, app := range apps {
-		appsRes = append(appsRes, AppBean{Id: app.Id, Name: app.AppName})
-	}
-	return appsRes, err
-}
-
-func (impl *PipelineBuilderImpl) FetchCDPipelineStrategy(appId int) (PipelineStrategiesResponse, error) {
-	pipelineStrategiesResponse := PipelineStrategiesResponse{}
-	chart, err := impl.chartRepository.FindLatestChartForAppByAppId(appId)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorf("invalid state", "err", err, "appId", appId)
-		return pipelineStrategiesResponse, err
-	}
-	if chart.Id == 0 {
-		return pipelineStrategiesResponse, fmt.Errorf("no chart configured")
-	}
-
-	//get global strategy for this chart
-	globalStrategies, err := impl.globalStrategyMetadataChartRefMappingRepository.GetByChartRefId(chart.ChartRefId)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("error in getting global strategies", "err", err)
-		return pipelineStrategiesResponse, err
-	} else if err == pg.ErrNoRows {
-		impl.logger.Infow("no strategies configured for chart", "chartRefId", chart.ChartRefId)
-		return pipelineStrategiesResponse, nil
-	}
-	pipelineOverride := chart.PipelineOverride
-	for _, globalStrategy := range globalStrategies {
-		pipelineStrategyJson, err := impl.filterDeploymentTemplate(globalStrategy.GlobalStrategyMetadata.Key, pipelineOverride)
-		if err != nil {
-			return pipelineStrategiesResponse, err
-		}
-		pipelineStrategy := PipelineStrategy{
-			DeploymentTemplate: globalStrategy.GlobalStrategyMetadata.Name,
-			Config:             []byte(pipelineStrategyJson),
-		}
-		pipelineStrategy.Default = globalStrategy.Default
-		pipelineStrategiesResponse.PipelineStrategy = append(pipelineStrategiesResponse.PipelineStrategy, pipelineStrategy)
-	}
-	return pipelineStrategiesResponse, nil
-}
-
-func (impl *PipelineBuilderImpl) FetchDefaultCDPipelineStrategy(appId int, envId int) (PipelineStrategy, error) {
-	pipelineStrategy := PipelineStrategy{}
-	cdPipelines, err := impl.ciCdPipelineOrchestrator.GetCdPipelinesForAppAndEnv(appId, envId)
-	if err != nil || (cdPipelines.Pipelines) == nil || len(cdPipelines.Pipelines) == 0 {
-		return pipelineStrategy, err
-	}
-	cdPipelineId := cdPipelines.Pipelines[0].Id
-
-	cdPipeline, err := impl.GetCdPipelineById(cdPipelineId)
-	if err != nil {
-		return pipelineStrategy, nil
-	}
-	pipelineStrategy.DeploymentTemplate = cdPipeline.DeploymentTemplate
-	pipelineStrategy.Default = true
-	return pipelineStrategy, nil
-}
-
 type PipelineStrategiesResponse struct {
 	PipelineStrategy []PipelineStrategy `json:"pipelineStrategy"`
 }
@@ -4237,290 +1804,6 @@ type PipelineStrategy struct {
 	DeploymentTemplate chartRepoRepository.DeploymentStrategy `json:"deploymentTemplate,omitempty"` //
 	Config             json.RawMessage                        `json:"config"`
 	Default            bool                                   `json:"default"`
-}
-
-func (impl *PipelineBuilderImpl) GetEnvironmentByCdPipelineId(pipelineId int) (int, error) {
-	dbPipeline, err := impl.pipelineRepository.FindById(pipelineId)
-	if err != nil || dbPipeline == nil {
-		impl.logger.Errorw("error in fetching pipeline", "err", err)
-		return 0, err
-	}
-	return dbPipeline.EnvironmentId, err
-}
-
-func (impl *PipelineBuilderImpl) GetCdPipelineById(pipelineId int) (cdPipeline *bean.CDPipelineConfigObject, err error) {
-	dbPipeline, err := impl.pipelineRepository.FindById(pipelineId)
-	if err != nil && errors.IsNotFound(err) {
-		impl.logger.Errorw("error in fetching pipeline", "err", err)
-		return cdPipeline, err
-	}
-	environment, err := impl.environmentRepository.FindById(dbPipeline.EnvironmentId)
-	if err != nil && errors.IsNotFound(err) {
-		impl.logger.Errorw("error in fetching pipeline", "err", err)
-		return cdPipeline, err
-	}
-	strategies, err := impl.pipelineConfigRepository.GetAllStrategyByPipelineId(dbPipeline.Id)
-	if err != nil && errors.IsNotFound(err) {
-		impl.logger.Errorw("error in fetching strategies", "err", err)
-		return cdPipeline, err
-	}
-	var strategiesBean []bean.Strategy
-	var deploymentTemplate chartRepoRepository.DeploymentStrategy
-	for _, item := range strategies {
-		strategiesBean = append(strategiesBean, bean.Strategy{
-			Config:             []byte(item.Config),
-			DeploymentTemplate: item.Strategy,
-			Default:            item.Default,
-		})
-
-		if item.Default {
-			deploymentTemplate = item.Strategy
-		}
-	}
-
-	preStage := bean.CdStage{}
-	if len(dbPipeline.PreStageConfig) > 0 {
-		preStage.Name = "Pre-Deployment"
-		preStage.Config = dbPipeline.PreStageConfig
-		preStage.TriggerType = dbPipeline.PreTriggerType
-	}
-	postStage := bean.CdStage{}
-	if len(dbPipeline.PostStageConfig) > 0 {
-		postStage.Name = "Post-Deployment"
-		postStage.Config = dbPipeline.PostStageConfig
-		postStage.TriggerType = dbPipeline.PostTriggerType
-	}
-
-	preStageConfigmapSecrets := bean.PreStageConfigMapSecretNames{}
-	postStageConfigmapSecrets := bean.PostStageConfigMapSecretNames{}
-
-	if dbPipeline.PreStageConfigMapSecretNames != "" {
-		err = json.Unmarshal([]byte(dbPipeline.PreStageConfigMapSecretNames), &preStageConfigmapSecrets)
-		if err != nil {
-			impl.logger.Error(err)
-			return nil, err
-		}
-	}
-	if dbPipeline.PostStageConfigMapSecretNames != "" {
-		err = json.Unmarshal([]byte(dbPipeline.PostStageConfigMapSecretNames), &postStageConfigmapSecrets)
-		if err != nil {
-			impl.logger.Error(err)
-			return nil, err
-		}
-	}
-	appWorkflowMapping, err := impl.appWorkflowRepository.FindWFCDMappingByCDPipelineId(pipelineId)
-	if err != nil {
-		return nil, err
-	}
-	cdPipeline = &bean.CDPipelineConfigObject{
-		Id:                            dbPipeline.Id,
-		Name:                          dbPipeline.Name,
-		EnvironmentId:                 dbPipeline.EnvironmentId,
-		EnvironmentName:               environment.Name,
-		CiPipelineId:                  dbPipeline.CiPipelineId,
-		DeploymentTemplate:            deploymentTemplate,
-		TriggerType:                   dbPipeline.TriggerType,
-		Strategies:                    strategiesBean,
-		PreStage:                      preStage,
-		PostStage:                     postStage,
-		PreStageConfigMapSecretNames:  preStageConfigmapSecrets,
-		PostStageConfigMapSecretNames: postStageConfigmapSecrets,
-		RunPreStageInEnv:              dbPipeline.RunPreStageInEnv,
-		RunPostStageInEnv:             dbPipeline.RunPostStageInEnv,
-		CdArgoSetup:                   environment.Cluster.CdArgoSetup,
-		ParentPipelineId:              appWorkflowMapping.ParentId,
-		ParentPipelineType:            appWorkflowMapping.ParentType,
-		DeploymentAppType:             dbPipeline.DeploymentAppType,
-		DeploymentAppCreated:          dbPipeline.DeploymentAppCreated,
-		IsVirtualEnvironment:          dbPipeline.Environment.IsVirtualEnvironment,
-	}
-	var preDeployStage *bean3.PipelineStageDto
-	var postDeployStage *bean3.PipelineStageDto
-	preDeployStage, postDeployStage, err = impl.pipelineStageService.GetCdPipelineStageDataDeepCopy(dbPipeline.Id)
-	if err != nil {
-		impl.logger.Errorw("error in getting pre/post-CD stage data", "err", err, "cdPipelineId", dbPipeline.Id)
-		return nil, err
-	}
-	cdPipeline.PreDeployStage = preDeployStage
-	cdPipeline.PostDeployStage = postDeployStage
-
-	return cdPipeline, err
-}
-
-func (impl *PipelineBuilderImpl) FindByIds(ids []*int) ([]*AppBean, error) {
-	var appsRes []*AppBean
-	apps, err := impl.appRepo.FindByIds(ids)
-	if err != nil {
-		impl.logger.Errorw("error while fetching app", "err", err)
-		return nil, err
-	}
-	for _, app := range apps {
-		appsRes = append(appsRes, &AppBean{Id: app.Id, Name: app.AppName, TeamId: app.TeamId})
-	}
-	return appsRes, err
-}
-
-func (impl *PipelineBuilderImpl) GetCiPipelineById(pipelineId int) (ciPipeline *bean.CiPipeline, err error) {
-	pipeline, err := impl.ciPipelineRepository.FindById(pipelineId)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching ci pipeline", "pipelineId", pipelineId, "err", err)
-		return nil, err
-	}
-	dockerArgs := make(map[string]string)
-	if len(pipeline.DockerArgs) > 0 {
-		err := json.Unmarshal([]byte(pipeline.DockerArgs), &dockerArgs)
-		if err != nil {
-			impl.logger.Warnw("error in unmarshal", "err", err)
-		}
-	}
-
-	if impl.ciConfig.ExternalCiWebhookUrl == "" {
-		hostUrl, err := impl.attributesService.GetByKey(attributes.HostUrlKey)
-		if err != nil {
-			impl.logger.Errorw("there is no external ci webhook url configured", "ci pipeline", pipeline)
-			return nil, err
-		}
-		if hostUrl != nil {
-			impl.ciConfig.ExternalCiWebhookUrl = fmt.Sprintf("%s/%s", hostUrl.Value, ExternalCiWebhookPath)
-		}
-	}
-
-	var externalCiConfig bean.ExternalCiConfig
-
-	ciPipelineScripts, err := impl.ciPipelineRepository.FindCiScriptsByCiPipelineId(pipeline.Id)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching ci scripts")
-		return nil, err
-	}
-
-	var beforeDockerBuildScripts []*bean.CiScript
-	var afterDockerBuildScripts []*bean.CiScript
-	for _, ciScript := range ciPipelineScripts {
-		ciScriptResp := &bean.CiScript{
-			Id:             ciScript.Id,
-			Index:          ciScript.Index,
-			Name:           ciScript.Name,
-			Script:         ciScript.Script,
-			OutputLocation: ciScript.OutputLocation,
-		}
-		if ciScript.Stage == BEFORE_DOCKER_BUILD {
-			beforeDockerBuildScripts = append(beforeDockerBuildScripts, ciScriptResp)
-		} else if ciScript.Stage == AFTER_DOCKER_BUILD {
-			afterDockerBuildScripts = append(afterDockerBuildScripts, ciScriptResp)
-		}
-	}
-	parentCiPipeline, err := impl.ciPipelineRepository.FindById(pipeline.ParentCiPipeline)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("err", err)
-		return nil, err
-	}
-	ciPipeline = &bean.CiPipeline{
-		Id:                       pipeline.Id,
-		Version:                  pipeline.Version,
-		Name:                     pipeline.Name,
-		Active:                   pipeline.Active,
-		Deleted:                  pipeline.Deleted,
-		DockerArgs:               dockerArgs,
-		IsManual:                 pipeline.IsManual,
-		IsExternal:               pipeline.IsExternal,
-		AppId:                    pipeline.AppId,
-		ParentCiPipeline:         pipeline.ParentCiPipeline,
-		ParentAppId:              parentCiPipeline.AppId,
-		ExternalCiConfig:         externalCiConfig,
-		BeforeDockerBuildScripts: beforeDockerBuildScripts,
-		AfterDockerBuildScripts:  afterDockerBuildScripts,
-		ScanEnabled:              pipeline.ScanEnabled,
-		IsDockerConfigOverridden: pipeline.IsDockerConfigOverridden,
-		PipelineType:             bean.PipelineType(pipeline.PipelineType),
-	}
-	ciEnvMapping, err := impl.ciPipelineRepository.FindCiEnvMappingByCiPipelineId(pipelineId)
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("error in fetching ci env mapping", "pipelineId", pipelineId, "err", err)
-		return nil, err
-	}
-	if ciEnvMapping.Id > 0 {
-		ciPipeline.EnvironmentId = ciEnvMapping.EnvironmentId
-	}
-	if !ciPipeline.IsExternal && ciPipeline.IsDockerConfigOverridden {
-		ciTemplateBean, err := impl.ciTemplateService.FindTemplateOverrideByCiPipelineId(ciPipeline.Id)
-		if err != nil {
-			return nil, err
-		}
-		templateOverride := ciTemplateBean.CiTemplateOverride
-		ciBuildConfig := ciTemplateBean.CiBuildConfig
-		ciPipeline.DockerConfigOverride = bean.DockerConfigOverride{
-			DockerRegistry:   templateOverride.DockerRegistryId,
-			DockerRepository: templateOverride.DockerRepository,
-			CiBuildConfig:    ciBuildConfig,
-			//DockerBuildConfig: &bean.DockerBuildConfig{
-			//	GitMaterialId:  templateOverride.GitMaterialId,
-			//	DockerfilePath: templateOverride.DockerfilePath,
-			//},
-		}
-	}
-	for _, material := range pipeline.CiPipelineMaterials {
-		if material == nil || material.GitMaterial == nil || !material.GitMaterial.Active {
-			continue
-		}
-		ciMaterial := &bean.CiMaterial{
-			Id:              material.Id,
-			CheckoutPath:    material.CheckoutPath,
-			Path:            material.Path,
-			ScmId:           material.ScmId,
-			GitMaterialId:   material.GitMaterialId,
-			GitMaterialName: material.GitMaterial.Name[strings.Index(material.GitMaterial.Name, "-")+1:],
-			ScmName:         material.ScmName,
-			ScmVersion:      material.ScmVersion,
-			IsRegex:         material.Regex != "",
-			Source:          &bean.SourceTypeConfig{Type: material.Type, Value: material.Value, Regex: material.Regex},
-		}
-		ciPipeline.CiMaterial = append(ciPipeline.CiMaterial, ciMaterial)
-	}
-
-	linkedCis, err := impl.ciPipelineRepository.FindByParentCiPipelineId(ciPipeline.Id)
-	if err != nil && !util.IsErrNoRows(err) {
-		return nil, err
-	}
-	ciPipeline.LinkedCount = len(linkedCis)
-
-	appWorkflowMappings, err := impl.appWorkflowRepository.FindWFCIMappingByCIPipelineId(ciPipeline.Id)
-	for _, mapping := range appWorkflowMappings {
-		//there will be only one active entry in db always
-		ciPipeline.AppWorkflowId = mapping.AppWorkflowId
-	}
-
-	//getting pre stage and post stage details
-	preStageDetail, postStageDetail, err := impl.pipelineStageService.GetCiPipelineStageData(ciPipeline.Id)
-	if err != nil {
-		impl.logger.Errorw("error in getting pre & post stage detail by ciPipelineId", "err", err, "ciPipelineId", ciPipeline.Id)
-		return nil, err
-	}
-	ciPipeline.PreBuildStage = preStageDetail
-	ciPipeline.PostBuildStage = postStageDetail
-	return ciPipeline, err
-}
-
-func (impl *PipelineBuilderImpl) FindAllMatchesByAppName(appName string, appType helper.AppType) ([]*AppBean, error) {
-	var appsRes []*AppBean
-	var apps []*app2.App
-	var err error
-	if len(appName) == 0 {
-		apps, err = impl.appRepo.FindAll()
-	} else {
-		apps, err = impl.appRepo.FindAllMatchesByAppName(appName, appType)
-	}
-	if err != nil {
-		impl.logger.Errorw("error while fetching app", "err", err)
-		return nil, err
-	}
-	for _, app := range apps {
-		name := app.AppName
-		if appType == helper.Job {
-			name = app.DisplayName
-		}
-		appsRes = append(appsRes, &AppBean{Id: app.Id, Name: name})
-	}
-	return appsRes, err
 }
 
 func (impl *PipelineBuilderImpl) updateGitRepoUrlInCharts(appId int, chartGitAttribute *util.ChartGitAttribute, userId int32) error {
@@ -4541,22 +1824,6 @@ func (impl *PipelineBuilderImpl) updateGitRepoUrlInCharts(appId int, chartGitAtt
 		}
 	}
 	return nil
-}
-
-func (impl *PipelineBuilderImpl) PerformBulkActionOnCdPipelines(dto *bean.CdBulkActionRequestDto, impactedPipelines []*pipelineConfig.Pipeline, ctx context.Context, dryRun bool, userId int32) ([]*bean.CdBulkActionResponseDto, error) {
-	switch dto.Action {
-	case bean.CD_BULK_DELETE:
-		deleteAction := bean.CASCADE_DELETE
-		if dto.ForceDelete {
-			deleteAction = bean.FORCE_DELETE
-		} else if !dto.CascadeDelete {
-			deleteAction = bean.NON_CASCADE_DELETE
-		}
-		bulkDeleteResp := impl.BulkDeleteCdPipelines(impactedPipelines, ctx, dryRun, deleteAction, userId)
-		return bulkDeleteResp, nil
-	default:
-		return nil, &util.ApiError{Code: "400", HttpStatusCode: 400, UserMessage: "this action is not supported"}
-	}
 }
 
 func (impl *PipelineBuilderImpl) BulkDeleteCdPipelines(impactedPipelines []*pipelineConfig.Pipeline, ctx context.Context, dryRun bool, deleteAction int, userId int32) []*bean.CdBulkActionResponseDto {
@@ -4582,47 +1849,6 @@ func (impl *PipelineBuilderImpl) BulkDeleteCdPipelines(impactedPipelines []*pipe
 	}
 	return respDtos
 
-}
-
-func (impl *PipelineBuilderImpl) GetBulkActionImpactedPipelines(dto *bean.CdBulkActionRequestDto) ([]*pipelineConfig.Pipeline, error) {
-	if len(dto.EnvIds) == 0 || (len(dto.AppIds) == 0 && len(dto.ProjectIds) == 0) {
-		//invalid payload, envIds are must and either of appIds or projectIds are must
-		return nil, &util.ApiError{Code: "400", HttpStatusCode: 400, UserMessage: "invalid payload, can not get pipelines for this filter"}
-	}
-	var pipelineIdsByAppLevel []int
-	var pipelineIdsByProjectLevel []int
-	var err error
-	if len(dto.AppIds) > 0 && len(dto.EnvIds) > 0 {
-		//getting pipeline IDs for app level deletion request
-		pipelineIdsByAppLevel, err = impl.pipelineRepository.FindIdsByAppIdsAndEnvironmentIds(dto.AppIds, dto.EnvIds)
-		if err != nil && err != pg.ErrNoRows {
-			impl.logger.Errorw("error in getting cd pipelines by appIds and envIds", "err", err)
-			return nil, err
-		}
-	}
-	if len(dto.ProjectIds) > 0 && len(dto.EnvIds) > 0 {
-		//getting pipeline IDs for project level deletion request
-		pipelineIdsByProjectLevel, err = impl.pipelineRepository.FindIdsByProjectIdsAndEnvironmentIds(dto.ProjectIds, dto.EnvIds)
-		if err != nil && err != pg.ErrNoRows {
-			impl.logger.Errorw("error in getting cd pipelines by projectIds and envIds", "err", err)
-			return nil, err
-		}
-	}
-	var pipelineIdsMerged []int
-	//it might be possible that pipelineIdsByAppLevel & pipelineIdsByProjectLevel have some same values
-	//we are still appending them to save operation cost of checking same ids as we will get pipelines from
-	//in clause which gives correct results even if some values are repeating
-	pipelineIdsMerged = append(pipelineIdsMerged, pipelineIdsByAppLevel...)
-	pipelineIdsMerged = append(pipelineIdsMerged, pipelineIdsByProjectLevel...)
-	var pipelines []*pipelineConfig.Pipeline
-	if len(pipelineIdsMerged) > 0 {
-		pipelines, err = impl.pipelineRepository.FindByIdsIn(pipelineIdsMerged)
-		if err != nil {
-			impl.logger.Errorw("error in getting cd pipelines by ids", "err", err, "ids", pipelineIdsMerged)
-			return nil, err
-		}
-	}
-	return pipelines, nil
 }
 
 func (impl *PipelineBuilderImpl) buildExternalCiWebhookSchema() map[string]interface{} {
@@ -4735,781 +1961,4 @@ func (impl *PipelineBuilderImpl) buildResponses() []bean.ResponseSchemaObject {
 	responseSchemaObjects = append(responseSchemaObjects, response400)
 	responseSchemaObjects = append(responseSchemaObjects, response401)
 	return responseSchemaObjects
-}
-
-func (impl *PipelineBuilderImpl) MarkGitOpsDevtronAppsDeletedWhereArgoAppIsDeleted(appId int, envId int, acdToken string, pipeline *pipelineConfig.Pipeline) (bool, error) {
-
-	acdAppFound := false
-	ctx := context.Background()
-	ctx = context.WithValue(ctx, "token", acdToken)
-	acdAppName := pipeline.DeploymentAppName
-	_, err := impl.application.Get(ctx, &application2.ApplicationQuery{Name: &acdAppName})
-	if err == nil {
-		// acd app is not yet deleted so return
-		acdAppFound = true
-		return acdAppFound, err
-	}
-	impl.logger.Warnw("app not found in argo, deleting from db ", "err", err)
-	//make call to delete it from pipeline DB because it's ACD counterpart is deleted
-	_, err = impl.DeleteCdPipeline(pipeline, context.Background(), bean.FORCE_DELETE, false, 1)
-	if err != nil {
-		impl.logger.Errorw("error in deleting cd pipeline", "err", err)
-		return acdAppFound, err
-	}
-	return acdAppFound, nil
-}
-
-func (impl PipelineBuilderImpl) GetCiPipelineByEnvironment(request resourceGroup2.ResourceGroupingRequest) ([]*bean.CiConfigRequest, error) {
-	ciPipelinesConfigByApps := make([]*bean.CiConfigRequest, 0)
-	_, span := otel.Tracer("orchestrator").Start(request.Ctx, "ciHandler.ResourceGroupingCiPipelinesAuthorization")
-	var cdPipelines []*pipelineConfig.Pipeline
-	var err error
-	if request.ResourceGroupId > 0 {
-		appIds, err := impl.resourceGroupService.GetResourceIdsByResourceGroupId(request.ResourceGroupId)
-		if err != nil {
-			return nil, err
-		}
-		//override appIds if already provided app group id in request.
-		request.ResourceIds = appIds
-	}
-	if len(request.ResourceIds) > 0 {
-		cdPipelines, err = impl.pipelineRepository.FindActiveByInFilter(request.ParentResourceId, request.ResourceIds)
-	} else {
-		cdPipelines, err = impl.pipelineRepository.FindActiveByEnvId(request.ParentResourceId)
-	}
-	if err != nil {
-		impl.logger.Errorw("error in fetching pipelines", "request", request, "err", err)
-		return nil, err
-	}
-
-	var appIds []int
-	ciPipelineIds := make([]int, 0)
-	cdPipelineIds := make([]int, 0)
-	for _, pipeline := range cdPipelines {
-		cdPipelineIds = append(cdPipelineIds, pipeline.Id)
-	}
-
-	//authorization block starts here
-	var appObjectArr []string
-	objects := impl.enforcerUtil.GetAppAndEnvObjectByDbPipeline(cdPipelines)
-	ciPipelineIds = []int{}
-	for _, object := range objects {
-		appObjectArr = append(appObjectArr, object[0])
-	}
-	appResults, _ := request.CheckAuthBatch(request.EmailId, appObjectArr, []string{})
-	for _, pipeline := range cdPipelines {
-		appObject := objects[pipeline.Id]
-		if !appResults[appObject[0]] {
-			//if user unauthorized, skip items
-			continue
-		}
-		appIds = append(appIds, pipeline.AppId)
-		ciPipelineIds = append(ciPipelineIds, pipeline.CiPipelineId)
-	}
-	//authorization block ends here
-	span.End()
-	if len(appIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no matching app found"}
-		return nil, err
-	}
-	if impl.ciConfig.ExternalCiWebhookUrl == "" {
-		hostUrl, err := impl.attributesService.GetByKey(attributes.HostUrlKey)
-		if err != nil {
-			return nil, err
-		}
-		if hostUrl != nil {
-			impl.ciConfig.ExternalCiWebhookUrl = fmt.Sprintf("%s/%s", hostUrl.Value, ExternalCiWebhookPath)
-		}
-	}
-
-	_, span = otel.Tracer("orchestrator").Start(request.Ctx, "ciHandler.GetCiTemplateVariables")
-	defer span.End()
-	ciPipelinesConfigMap, err := impl.getCiTemplateVariablesByAppIds(appIds)
-	if err != nil {
-		impl.logger.Debugw("error in fetching ci pipeline", "appIds", appIds, "err", err)
-		return nil, err
-	}
-
-	ciPipelineByApp := make(map[int][]*pipelineConfig.CiPipeline)
-	_, span = otel.Tracer("orchestrator").Start(request.Ctx, "ciHandler.FindByAppIds")
-	ciPipelines, err := impl.ciPipelineRepository.FindByAppIds(appIds)
-	span.End()
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching ci pipeline", "appIds", appIds, "err", err)
-		return nil, err
-	}
-	parentCiPipelineIds := make([]int, 0)
-	for _, ciPipeline := range ciPipelines {
-		ciPipelineByApp[ciPipeline.AppId] = append(ciPipelineByApp[ciPipeline.AppId], ciPipeline)
-		if ciPipeline.ParentCiPipeline > 0 && ciPipeline.IsExternal {
-			parentCiPipelineIds = append(parentCiPipelineIds, ciPipeline.ParentCiPipeline)
-		}
-	}
-	pipelineIdVsAppId, err := impl.ciPipelineRepository.FindAppIdsForCiPipelineIds(parentCiPipelineIds)
-	if err != nil {
-		impl.logger.Errorw("error occurred while fetching appIds for pipelineIds", "parentCiPipelineIds", parentCiPipelineIds, "err", err)
-		return nil, err
-	}
-
-	if len(ciPipelineIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no matching ci pipeline found"}
-		return nil, err
-	}
-	linkedCiPipelinesMap := make(map[int][]*pipelineConfig.CiPipeline)
-	_, span = otel.Tracer("orchestrator").Start(request.Ctx, "ciHandler.FindByParentCiPipelineIds")
-	linkedCiPipelines, err := impl.ciPipelineRepository.FindByParentCiPipelineIds(ciPipelineIds)
-	span.End()
-	if err != nil && !util.IsErrNoRows(err) {
-		return nil, err
-	}
-	for _, linkedCiPipeline := range linkedCiPipelines {
-		linkedCiPipelinesMap[linkedCiPipeline.ParentCiPipeline] = append(linkedCiPipelinesMap[linkedCiPipeline.Id], linkedCiPipeline)
-	}
-
-	_, span = otel.Tracer("orchestrator").Start(request.Ctx, "ciHandler.FindTemplateOverrideByCiPipelineIds")
-	ciTemplateBeanOverrides, err := impl.ciTemplateService.FindTemplateOverrideByCiPipelineIds(ciPipelineIds)
-	span.End()
-	if err != nil {
-		impl.logger.Errorw("error in fetching templates override", "appIds", appIds, "err", err)
-		return nil, err
-	}
-	ciOverrideTemplateMap := make(map[int]*bean3.CiTemplateBean)
-	for _, templateBeanOverride := range ciTemplateBeanOverrides {
-		ciTemplateOverride := templateBeanOverride.CiTemplateOverride
-		ciOverrideTemplateMap[ciTemplateOverride.CiPipelineId] = templateBeanOverride
-	}
-
-	var externalCiConfig bean.ExternalCiConfig
-	//var parentCiPipelineIds []int
-	for appId, ciPipelinesConfigByApp := range ciPipelinesConfigMap {
-		var ciPipelineResp []*bean.CiPipeline
-
-		ciPipelines := ciPipelineByApp[appId]
-		for _, pipeline := range ciPipelines {
-			dockerArgs := make(map[string]string)
-			if len(pipeline.DockerArgs) > 0 {
-				err := json.Unmarshal([]byte(pipeline.DockerArgs), &dockerArgs)
-				if err != nil {
-					impl.logger.Warnw("error in unmarshal", "err", err)
-				}
-			}
-			parentCiPipelineId := pipeline.ParentCiPipeline
-			ciPipeline := &bean.CiPipeline{
-				Id:                       pipeline.Id,
-				Version:                  pipeline.Version,
-				Name:                     pipeline.Name,
-				Active:                   pipeline.Active,
-				Deleted:                  pipeline.Deleted,
-				DockerArgs:               dockerArgs,
-				IsManual:                 pipeline.IsManual,
-				IsExternal:               pipeline.IsExternal,
-				ParentCiPipeline:         parentCiPipelineId,
-				ExternalCiConfig:         externalCiConfig,
-				ScanEnabled:              pipeline.ScanEnabled,
-				IsDockerConfigOverridden: pipeline.IsDockerConfigOverridden,
-				PipelineType:             bean.PipelineType(pipeline.PipelineType),
-			}
-			parentPipelineAppId, ok := pipelineIdVsAppId[parentCiPipelineId]
-			if ok {
-				ciPipeline.ParentAppId = parentPipelineAppId
-			}
-			if ciTemplateBean, ok := ciOverrideTemplateMap[pipeline.Id]; ok {
-				templateOverride := ciTemplateBean.CiTemplateOverride
-				ciPipeline.DockerConfigOverride = bean.DockerConfigOverride{
-					DockerRegistry:   templateOverride.DockerRegistryId,
-					DockerRepository: templateOverride.DockerRepository,
-					CiBuildConfig:    ciTemplateBean.CiBuildConfig,
-				}
-			}
-
-			//this will build ci materials for each ci pipeline
-			for _, material := range pipeline.CiPipelineMaterials {
-				// ignore those materials which have inactive git material
-				if material == nil || material.GitMaterial == nil || !material.GitMaterial.Active {
-					continue
-				}
-				ciMaterial := &bean.CiMaterial{
-					Id:              material.Id,
-					CheckoutPath:    material.CheckoutPath,
-					Path:            material.Path,
-					ScmId:           material.ScmId,
-					GitMaterialId:   material.GitMaterialId,
-					GitMaterialName: material.GitMaterial.Name[strings.Index(material.GitMaterial.Name, "-")+1:],
-					ScmName:         material.ScmName,
-					ScmVersion:      material.ScmVersion,
-					IsRegex:         material.Regex != "",
-					Source:          &bean.SourceTypeConfig{Type: material.Type, Value: material.Value, Regex: material.Regex},
-				}
-				ciPipeline.CiMaterial = append(ciPipeline.CiMaterial, ciMaterial)
-			}
-
-			//this will count the length of child ci pipelines, of each ci pipeline
-			linkedCi := linkedCiPipelinesMap[pipeline.Id]
-			ciPipeline.LinkedCount = len(linkedCi)
-			ciPipelineResp = append(ciPipelineResp, ciPipeline)
-
-			//this will use for fetch the parent ci pipeline app id, of each ci pipeline
-			//parentCiPipelineIds = append(parentCiPipelineIds, pipeline.ParentCiPipeline)
-		}
-		ciPipelinesConfigByApp.CiPipelines = ciPipelineResp
-		ciPipelinesConfigByApp.CiGitMaterialId = ciPipelinesConfigByApp.CiBuildConfig.GitMaterialId
-		ciPipelinesConfigByApps = append(ciPipelinesConfigByApps, ciPipelinesConfigByApp)
-	}
-
-	return ciPipelinesConfigByApps, err
-}
-
-func (impl PipelineBuilderImpl) GetCiPipelineByEnvironmentMin(request resourceGroup2.ResourceGroupingRequest) ([]*bean.CiPipelineMinResponse, error) {
-	results := make([]*bean.CiPipelineMinResponse, 0)
-	var cdPipelines []*pipelineConfig.Pipeline
-	var err error
-	if request.ResourceGroupId > 0 {
-		appIds, err := impl.resourceGroupService.GetResourceIdsByResourceGroupId(request.ResourceGroupId)
-		if err != nil {
-			return results, err
-		}
-		//override appIds if already provided app group id in request.
-		request.ResourceIds = appIds
-	}
-	if len(request.ResourceIds) > 0 {
-		cdPipelines, err = impl.pipelineRepository.FindActiveByInFilter(request.ParentResourceId, request.ResourceIds)
-	} else {
-		cdPipelines, err = impl.pipelineRepository.FindActiveByEnvId(request.ParentResourceId)
-	}
-	if err != nil {
-		impl.logger.Errorw("error in fetching pipelines", "request", request, "err", err)
-		return results, err
-	}
-	foundAppIds := make([]int, 0)
-	for _, pipeline := range cdPipelines {
-		foundAppIds = append(foundAppIds, pipeline.AppId)
-	}
-	if len(foundAppIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no matching pipeline found"}
-		return nil, err
-	}
-	ciPipelines, err := impl.ciPipelineRepository.FindByAppIds(foundAppIds)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching ci pipeline", "err", err)
-		return nil, err
-	}
-	ciPipelineByApp := make(map[int]*pipelineConfig.CiPipeline)
-	parentCiPipelineIds := make([]int, 0)
-	for _, ciPipeline := range ciPipelines {
-		ciPipelineByApp[ciPipeline.Id] = ciPipeline
-		if ciPipeline.ParentCiPipeline > 0 && ciPipeline.IsExternal {
-			parentCiPipelineIds = append(parentCiPipelineIds, ciPipeline.ParentCiPipeline)
-		}
-	}
-	pipelineIdVsAppId, err := impl.ciPipelineRepository.FindAppIdsForCiPipelineIds(parentCiPipelineIds)
-	if err != nil {
-		impl.logger.Errorw("error occurred while fetching appIds for pipelineIds", "parentCiPipelineIds", parentCiPipelineIds, "err", err)
-		return nil, err
-	}
-
-	//authorization block starts here
-	var appObjectArr []string
-	objects := impl.enforcerUtil.GetAppAndEnvObjectByDbPipeline(cdPipelines)
-	for _, object := range objects {
-		appObjectArr = append(appObjectArr, object[0])
-	}
-	appResults, _ := request.CheckAuthBatch(request.EmailId, appObjectArr, []string{})
-	authorizedIds := make([]int, 0)
-	for _, pipeline := range cdPipelines {
-		appObject := objects[pipeline.Id]
-		if !appResults[appObject[0]] {
-			//if user unauthorized, skip items
-			continue
-		}
-		if pipeline.CiPipelineId == 0 {
-			//skip for external ci
-			continue
-		}
-		ciPipeline := ciPipelineByApp[pipeline.CiPipelineId]
-		parentAppId := pipelineIdVsAppId[ciPipeline.ParentCiPipeline]
-		result := &bean.CiPipelineMinResponse{
-			Id:               pipeline.CiPipelineId,
-			AppId:            pipeline.AppId,
-			AppName:          pipeline.App.AppName,
-			ParentCiPipeline: ciPipeline.ParentCiPipeline,
-			ParentAppId:      parentAppId,
-			PipelineType:     ciPipeline.PipelineType,
-		}
-		results = append(results, result)
-		authorizedIds = append(authorizedIds, pipeline.CiPipelineId)
-	}
-	//authorization block ends here
-
-	return results, err
-}
-
-func (impl PipelineBuilderImpl) GetCdPipelinesByEnvironment(request resourceGroup2.ResourceGroupingRequest) (cdPipelines *bean.CdPipelines, err error) {
-	_, span := otel.Tracer("orchestrator").Start(request.Ctx, "cdHandler.authorizationCdPipelinesForResourceGrouping")
-	if request.ResourceGroupId > 0 {
-		appIds, err := impl.resourceGroupService.GetResourceIdsByResourceGroupId(request.ResourceGroupId)
-		if err != nil {
-			return nil, err
-		}
-		//override appIds if already provided app group id in request.
-		request.ResourceIds = appIds
-	}
-	cdPipelines, err = impl.ciCdPipelineOrchestrator.GetCdPipelinesForEnv(request.ParentResourceId, request.ResourceIds)
-	if err != nil {
-		impl.logger.Errorw("error in fetching pipeline", "err", err)
-		return cdPipelines, err
-	}
-	pipelineIds := make([]int, 0)
-	for _, pipeline := range cdPipelines.Pipelines {
-		pipelineIds = append(pipelineIds, pipeline.Id)
-	}
-	if len(pipelineIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no matching pipeline found"}
-		return cdPipelines, err
-	}
-	//authorization block starts here
-	var appObjectArr []string
-	var envObjectArr []string
-	objects := impl.enforcerUtil.GetAppAndEnvObjectByPipeline(cdPipelines.Pipelines)
-	pipelineIds = []int{}
-	for _, object := range objects {
-		appObjectArr = append(appObjectArr, object[0])
-		envObjectArr = append(envObjectArr, object[1])
-	}
-	appResults, envResults := request.CheckAuthBatch(request.EmailId, appObjectArr, envObjectArr)
-	//authorization block ends here
-	span.End()
-	var pipelines []*bean.CDPipelineConfigObject
-	authorizedPipelines := make(map[int]*bean.CDPipelineConfigObject)
-	for _, dbPipeline := range cdPipelines.Pipelines {
-		appObject := objects[dbPipeline.Id][0]
-		envObject := objects[dbPipeline.Id][1]
-		if !(appResults[appObject] && envResults[envObject]) {
-			//if user unauthorized, skip items
-			continue
-		}
-		pipelineIds = append(pipelineIds, dbPipeline.Id)
-		authorizedPipelines[dbPipeline.Id] = dbPipeline
-	}
-
-	pipelineDeploymentTemplate := make(map[int]chartRepoRepository.DeploymentStrategy)
-	pipelineWorkflowMapping := make(map[int]*appWorkflow.AppWorkflowMapping)
-	if len(pipelineIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no authorized pipeline found"}
-		return cdPipelines, err
-	}
-	_, span = otel.Tracer("orchestrator").Start(request.Ctx, "cdHandler.GetAllStrategyByPipelineIds")
-	strategies, err := impl.pipelineConfigRepository.GetAllStrategyByPipelineIds(pipelineIds)
-	span.End()
-	if err != nil {
-		impl.logger.Errorw("error in fetching strategies", "err", err)
-		return cdPipelines, err
-	}
-	for _, item := range strategies {
-		if item.Default {
-			pipelineDeploymentTemplate[item.PipelineId] = item.Strategy
-		}
-	}
-	_, span = otel.Tracer("orchestrator").Start(request.Ctx, "cdHandler.FindByCDPipelineIds")
-	appWorkflowMappings, err := impl.appWorkflowRepository.FindByCDPipelineIds(pipelineIds)
-	span.End()
-	if err != nil {
-		impl.logger.Errorw("error in fetching workflows", "err", err)
-		return nil, err
-	}
-	for _, item := range appWorkflowMappings {
-		pipelineWorkflowMapping[item.ComponentId] = item
-	}
-
-	for _, dbPipeline := range authorizedPipelines {
-		pipeline := &bean.CDPipelineConfigObject{
-			Id:                            dbPipeline.Id,
-			Name:                          dbPipeline.Name,
-			EnvironmentId:                 dbPipeline.EnvironmentId,
-			EnvironmentName:               dbPipeline.EnvironmentName,
-			CiPipelineId:                  dbPipeline.CiPipelineId,
-			DeploymentTemplate:            pipelineDeploymentTemplate[dbPipeline.Id],
-			TriggerType:                   dbPipeline.TriggerType,
-			PreStage:                      dbPipeline.PreStage,
-			PostStage:                     dbPipeline.PostStage,
-			PreStageConfigMapSecretNames:  dbPipeline.PreStageConfigMapSecretNames,
-			PostStageConfigMapSecretNames: dbPipeline.PostStageConfigMapSecretNames,
-			RunPreStageInEnv:              dbPipeline.RunPreStageInEnv,
-			RunPostStageInEnv:             dbPipeline.RunPostStageInEnv,
-			DeploymentAppType:             dbPipeline.DeploymentAppType,
-			ParentPipelineType:            pipelineWorkflowMapping[dbPipeline.Id].ParentType,
-			ParentPipelineId:              pipelineWorkflowMapping[dbPipeline.Id].ParentId,
-			AppName:                       dbPipeline.AppName,
-			AppId:                         dbPipeline.AppId,
-			IsVirtualEnvironment:          dbPipeline.IsVirtualEnvironment,
-			PreDeployStage:                dbPipeline.PreDeployStage,
-			PostDeployStage:               dbPipeline.PostDeployStage,
-		}
-		pipelines = append(pipelines, pipeline)
-	}
-	cdPipelines.Pipelines = pipelines
-	return cdPipelines, err
-}
-
-func (impl PipelineBuilderImpl) GetCdPipelinesByEnvironmentMin(request resourceGroup2.ResourceGroupingRequest) (cdPipelines []*bean.CDPipelineConfigObject, err error) {
-	_, span := otel.Tracer("orchestrator").Start(request.Ctx, "cdHandler.authorizationCdPipelinesForResourceGrouping")
-	if request.ResourceGroupId > 0 {
-		appIds, err := impl.resourceGroupService.GetResourceIdsByResourceGroupId(request.ResourceGroupId)
-		if err != nil {
-			return cdPipelines, err
-		}
-		//override appIds if already provided app group id in request.
-		request.ResourceIds = appIds
-	}
-	var pipelines []*pipelineConfig.Pipeline
-	if len(request.ResourceIds) > 0 {
-		pipelines, err = impl.pipelineRepository.FindActiveByInFilter(request.ParentResourceId, request.ResourceIds)
-	} else {
-		pipelines, err = impl.pipelineRepository.FindActiveByEnvId(request.ParentResourceId)
-	}
-	if err != nil {
-		impl.logger.Errorw("error in fetching pipelines", "request", request, "err", err)
-		return cdPipelines, err
-	}
-	//authorization block starts here
-	var appObjectArr []string
-	var envObjectArr []string
-	objects := impl.enforcerUtil.GetAppAndEnvObjectByDbPipeline(pipelines)
-	for _, object := range objects {
-		appObjectArr = append(appObjectArr, object[0])
-		envObjectArr = append(envObjectArr, object[1])
-	}
-	appResults, envResults := request.CheckAuthBatch(request.EmailId, appObjectArr, envObjectArr)
-	//authorization block ends here
-	span.End()
-	for _, dbPipeline := range pipelines {
-		appObject := objects[dbPipeline.Id][0]
-		envObject := objects[dbPipeline.Id][1]
-		if !(appResults[appObject] && envResults[envObject]) {
-			//if user unauthorized, skip items
-			continue
-		}
-		pcObject := &bean.CDPipelineConfigObject{
-			AppId:                dbPipeline.AppId,
-			AppName:              dbPipeline.App.AppName,
-			EnvironmentId:        dbPipeline.EnvironmentId,
-			Id:                   dbPipeline.Id,
-			DeploymentAppType:    dbPipeline.DeploymentAppType,
-			IsVirtualEnvironment: dbPipeline.Environment.IsVirtualEnvironment,
-		}
-		cdPipelines = append(cdPipelines, pcObject)
-	}
-	return cdPipelines, err
-}
-
-func (impl PipelineBuilderImpl) GetExternalCiByEnvironment(request resourceGroup2.ResourceGroupingRequest) (ciConfig []*bean.ExternalCiConfig, err error) {
-	_, span := otel.Tracer("orchestrator").Start(request.Ctx, "ciHandler.authorizationExternalCiForResourceGrouping")
-	externalCiConfigs := make([]*bean.ExternalCiConfig, 0)
-	var cdPipelines []*pipelineConfig.Pipeline
-	if request.ResourceGroupId > 0 {
-		appIds, err := impl.resourceGroupService.GetResourceIdsByResourceGroupId(request.ResourceGroupId)
-		if err != nil {
-			return nil, err
-		}
-		//override appIds if already provided app group id in request.
-		request.ResourceIds = appIds
-	}
-	if len(request.ResourceIds) > 0 {
-		cdPipelines, err = impl.pipelineRepository.FindActiveByInFilter(request.ParentResourceId, request.ResourceIds)
-	} else {
-		cdPipelines, err = impl.pipelineRepository.FindActiveByEnvId(request.ParentResourceId)
-	}
-	if err != nil {
-		impl.logger.Errorw("error in fetching pipelines", "request", request, "err", err)
-		return nil, err
-	}
-
-	var appIds []int
-	//authorization block starts here
-	var appObjectArr []string
-	objects := impl.enforcerUtil.GetAppAndEnvObjectByDbPipeline(cdPipelines)
-	for _, object := range objects {
-		appObjectArr = append(appObjectArr, object[0])
-	}
-	appResults, _ := request.CheckAuthBatch(request.EmailId, appObjectArr, []string{})
-	for _, pipeline := range cdPipelines {
-		appObject := objects[pipeline.Id]
-		if !appResults[appObject[0]] {
-			//if user unauthorized, skip items
-			continue
-		}
-		//add only those who have external ci
-		if pipeline.CiPipelineId == 0 {
-			appIds = append(appIds, pipeline.AppId)
-		}
-	}
-
-	//authorization block ends here
-	span.End()
-
-	if len(appIds) == 0 {
-		impl.logger.Warnw("there is no app id found for fetching external ci pipelines", "request", request)
-		return externalCiConfigs, nil
-	}
-	_, span = otel.Tracer("orchestrator").Start(request.Ctx, "ciHandler.FindExternalCiByAppIds")
-	externalCiPipelines, err := impl.ciPipelineRepository.FindExternalCiByAppIds(appIds)
-	span.End()
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching external ci", "request", request, "err", err)
-		return nil, err
-	}
-	hostUrl, err := impl.attributesService.GetByKey(attributes.HostUrlKey)
-	if err != nil {
-		impl.logger.Errorw("error in fetching external ci", "request", request, "err", err)
-		return nil, err
-	}
-	if hostUrl != nil {
-		impl.ciConfig.ExternalCiWebhookUrl = fmt.Sprintf("%s/%s", hostUrl.Value, ExternalCiWebhookPath)
-	}
-
-	var externalCiPipelineIds []int
-	appWorkflowMappingsMap := make(map[int][]*appWorkflow.AppWorkflowMapping)
-
-	for _, externalCiPipeline := range externalCiPipelines {
-		externalCiPipelineIds = append(externalCiPipelineIds, externalCiPipeline.Id)
-	}
-	if len(externalCiPipelineIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no external ci pipeline found"}
-		return externalCiConfigs, err
-	}
-	appWorkflowMappings, err := impl.appWorkflowRepository.FindWFCDMappingByExternalCiIdByIdsIn(externalCiPipelineIds)
-	if err != nil {
-		impl.logger.Errorw("Error in fetching app workflow mapping for CD pipeline by external CI ID", "err", err)
-		return nil, err
-	}
-
-	CDPipelineMap := make(map[int]*pipelineConfig.Pipeline)
-	appIdMap := make(map[int]*app2.App)
-	var componentIds []int
-	for _, appWorkflowMapping := range appWorkflowMappings {
-		appWorkflowMappingsMap[appWorkflowMapping.ParentId] = append(appWorkflowMappingsMap[appWorkflowMapping.ParentId], appWorkflowMapping)
-		componentIds = append(componentIds, appWorkflowMapping.ComponentId)
-	}
-	if len(componentIds) == 0 {
-		return nil, err
-	}
-	cdPipelines, err = impl.pipelineRepository.FindAppAndEnvironmentAndProjectByPipelineIds(componentIds)
-	if err != nil && !util.IsErrNoRows(err) {
-		impl.logger.Errorw("error in fetching external ci", "request", request, "err", err)
-		return nil, err
-	}
-	for _, pipeline := range cdPipelines {
-		CDPipelineMap[pipeline.Id] = pipeline
-		appIds = append(appIds, pipeline.AppId)
-	}
-	if len(appIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no matching apps found"}
-		return nil, err
-	}
-	apps, err := impl.appRepo.FindAppAndProjectByIdsIn(appIds)
-	for _, app := range apps {
-		appIdMap[app.Id] = app
-	}
-
-	_, span = otel.Tracer("orchestrator").Start(request.Ctx, "ciHandler.FindAppAndEnvironmentAndProjectByPipelineIds")
-	for _, externalCiPipeline := range externalCiPipelines {
-		externalCiConfig := &bean.ExternalCiConfig{
-			Id:         externalCiPipeline.Id,
-			WebhookUrl: fmt.Sprintf("%s/%d", impl.ciConfig.ExternalCiWebhookUrl, externalCiPipeline.Id),
-			Payload:    impl.ciConfig.ExternalCiPayload,
-			AccessKey:  "",
-		}
-
-		if _, ok := appWorkflowMappingsMap[externalCiPipeline.Id]; !ok {
-			return nil, errors.New("Error in fetching app workflow mapping for cd pipeline by parent id")
-		}
-		appWorkflowMappings := appWorkflowMappingsMap[externalCiPipeline.Id]
-		roleData := make(map[string]interface{})
-		for _, appWorkflowMapping := range appWorkflowMappings {
-			if _, ok := CDPipelineMap[appWorkflowMapping.ComponentId]; !ok {
-				impl.logger.Errorw("error in getting cd pipeline data for workflow", "app workflow id", appWorkflowMapping.ComponentId, "err", err)
-				return nil, errors.New("error in getting cd pipeline data for workflow")
-			}
-			cdPipeline := CDPipelineMap[appWorkflowMapping.ComponentId]
-			if _, ok := roleData[teamIdKey]; !ok {
-				if _, ok := appIdMap[cdPipeline.AppId]; !ok {
-					impl.logger.Errorw("error in getting app data for pipeline", "app id", cdPipeline.AppId, "err", err)
-					return nil, errors.New("error in getting app data for pipeline")
-				}
-				app := appIdMap[cdPipeline.AppId]
-				roleData[teamIdKey] = app.TeamId
-				roleData[teamNameKey] = app.Team.Name
-				roleData[appIdKey] = cdPipeline.AppId
-				roleData[appNameKey] = cdPipeline.App.AppName
-			}
-			if _, ok := roleData[environmentNameKey]; !ok {
-				roleData[environmentNameKey] = cdPipeline.Environment.Name
-			} else {
-				roleData[environmentNameKey] = fmt.Sprintf("%s,%s", roleData[environmentNameKey], cdPipeline.Environment.Name)
-			}
-			if _, ok := roleData[environmentIdentifierKey]; !ok {
-				roleData[environmentIdentifierKey] = cdPipeline.Environment.EnvironmentIdentifier
-			} else {
-				roleData[environmentIdentifierKey] = fmt.Sprintf("%s,%s", roleData[environmentIdentifierKey], cdPipeline.Environment.EnvironmentIdentifier)
-			}
-		}
-
-		externalCiConfig.ExternalCiConfigRole = bean.ExternalCiConfigRole{
-			ProjectId:             roleData[teamIdKey].(int),
-			ProjectName:           roleData[teamNameKey].(string),
-			AppId:                 roleData[appIdKey].(int),
-			AppName:               roleData[appNameKey].(string),
-			EnvironmentName:       roleData[environmentNameKey].(string),
-			EnvironmentIdentifier: roleData[environmentIdentifierKey].(string),
-			Role:                  "Build and deploy",
-		}
-		externalCiConfigs = append(externalCiConfigs, externalCiConfig)
-	}
-	span.End()
-	//--------pipeline population end
-	return externalCiConfigs, err
-}
-
-func (impl PipelineBuilderImpl) GetEnvironmentListForAutocompleteFilter(envName string, clusterIds []int, offset int, size int, emailId string, checkAuthBatch func(emailId string, appObject []string, envObject []string) (map[string]bool, map[string]bool), ctx context.Context) (*cluster.ResourceGroupingResponse, error) {
-	result := &cluster.ResourceGroupingResponse{}
-	var models []*repository2.Environment
-	var beans []cluster.EnvironmentBean
-	var err error
-	if len(envName) > 0 && len(clusterIds) > 0 {
-		models, err = impl.environmentRepository.FindByEnvNameAndClusterIds(envName, clusterIds)
-	} else if len(clusterIds) > 0 {
-		models, err = impl.environmentRepository.FindByClusterIdsWithFilter(clusterIds)
-	} else if len(envName) > 0 {
-		models, err = impl.environmentRepository.FindByEnvName(envName)
-	} else {
-		models, err = impl.environmentRepository.FindAllActiveWithFilter()
-	}
-	if err != nil && err != pg.ErrNoRows {
-		impl.logger.Errorw("error in fetching environment", "err", err)
-		return result, err
-	}
-	var envIds []int
-	for _, model := range models {
-		envIds = append(envIds, model.Id)
-	}
-	if len(envIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no matching environment found"}
-		return nil, err
-	}
-	_, span := otel.Tracer("orchestrator").Start(ctx, "pipelineBuilder.FindActiveByEnvIds")
-	cdPipelines, err := impl.pipelineRepository.FindActiveByEnvIds(envIds)
-	span.End()
-	if err != nil && err != pg.ErrNoRows {
-		return result, err
-	}
-	pipelineIds := make([]int, 0)
-	for _, pipeline := range cdPipelines {
-		pipelineIds = append(pipelineIds, pipeline.Id)
-	}
-	if len(pipelineIds) == 0 {
-		err = &util.ApiError{Code: "404", HttpStatusCode: 200, UserMessage: "no matching pipeline found"}
-		return nil, err
-	}
-	//authorization block starts here
-	var appObjectArr []string
-	var envObjectArr []string
-	_, span = otel.Tracer("orchestrator").Start(ctx, "pipelineBuilder.GetAppAndEnvObjectByPipelineIds")
-	objects := impl.enforcerUtil.GetAppAndEnvObjectByPipelineIds(pipelineIds)
-	span.End()
-	pipelineIds = []int{}
-	for _, object := range objects {
-		appObjectArr = append(appObjectArr, object[0])
-		envObjectArr = append(envObjectArr, object[1])
-	}
-	_, span = otel.Tracer("orchestrator").Start(ctx, "pipelineBuilder.checkAuthBatch")
-	appResults, envResults := checkAuthBatch(emailId, appObjectArr, envObjectArr)
-	span.End()
-	//authorization block ends here
-
-	pipelinesMap := make(map[int][]*pipelineConfig.Pipeline)
-	for _, pipeline := range cdPipelines {
-		appObject := objects[pipeline.Id][0]
-		envObject := objects[pipeline.Id][1]
-		if !(appResults[appObject] && envResults[envObject]) {
-			//if user unauthorized, skip items
-			continue
-		}
-		pipelinesMap[pipeline.EnvironmentId] = append(pipelinesMap[pipeline.EnvironmentId], pipeline)
-	}
-	for _, model := range models {
-		environment := cluster.EnvironmentBean{
-			Id:                    model.Id,
-			Environment:           model.Name,
-			Namespace:             model.Namespace,
-			CdArgoSetup:           model.Cluster.CdArgoSetup,
-			EnvironmentIdentifier: model.EnvironmentIdentifier,
-			ClusterName:           model.Cluster.ClusterName,
-			IsVirtualEnvironment:  model.IsVirtualEnvironment,
-		}
-
-		//authorization block starts here
-		appCount := 0
-		envPipelines := pipelinesMap[model.Id]
-		if _, ok := pipelinesMap[model.Id]; ok {
-			appCount = len(envPipelines)
-		}
-		environment.AppCount = appCount
-		beans = append(beans, environment)
-	}
-
-	envCount := len(beans)
-	// Apply pagination
-	if size > 0 {
-		if offset+size <= len(beans) {
-			beans = beans[offset : offset+size]
-		} else {
-			beans = beans[offset:]
-		}
-	}
-	result.EnvList = beans
-	result.EnvCount = envCount
-	return result, nil
-}
-
-func (impl PipelineBuilderImpl) GetAppListForEnvironment(request resourceGroup2.ResourceGroupingRequest) ([]*AppBean, error) {
-	var applicationList []*AppBean
-	var cdPipelines []*pipelineConfig.Pipeline
-	var err error
-	if request.ResourceGroupId > 0 {
-		appIds, err := impl.resourceGroupService.GetResourceIdsByResourceGroupId(request.ResourceGroupId)
-		if err != nil {
-			return nil, err
-		}
-		//override appIds if already provided app group id in request.
-		request.ResourceIds = appIds
-	}
-	if len(request.ResourceIds) > 0 {
-		cdPipelines, err = impl.pipelineRepository.FindActiveByInFilter(request.ParentResourceId, request.ResourceIds)
-	} else {
-		cdPipelines, err = impl.pipelineRepository.FindActiveByEnvId(request.ParentResourceId)
-	}
-	if err != nil {
-		impl.logger.Errorw("error in fetching pipelines", "request", request, "err", err)
-		return nil, err
-	}
-	if len(cdPipelines) == 0 {
-		return applicationList, nil
-	}
-	var appObjectArr []string
-	var envObjectArr []string
-	objects := impl.enforcerUtil.GetAppAndEnvObjectByDbPipeline(cdPipelines)
-	for _, object := range objects {
-		appObjectArr = append(appObjectArr, object[0])
-		envObjectArr = append(envObjectArr, object[1])
-	}
-	appResults, envResults := request.CheckAuthBatch(request.EmailId, appObjectArr, envObjectArr)
-	for _, pipeline := range cdPipelines {
-		appObject := objects[pipeline.Id][0]
-		envObject := objects[pipeline.Id][1]
-		if !(appResults[appObject] && envResults[envObject]) {
-			//if user unauthorized, skip items
-			continue
-		}
-		applicationList = append(applicationList, &AppBean{Id: pipeline.AppId, Name: pipeline.App.AppName})
-	}
-	return applicationList, err
 }
